[Migrate]Move model-specific helpers out of serving_speech (#6697)
