+++ tests/entrypoints/openai_api/test_serving_speech.py
@@ -53,7 +53,7 @@ from vllm_omni.entrypoints.openai.tts_adapters.base import (
 )
 from vllm_omni.entrypoints.openai.tts_adapters.capabilities import load_supported_speakers
 from vllm_omni.entrypoints.openai.tts_adapters.ming_tts import MingTTSAdapter
-from vllm_omni.entrypoints.openai.tts_adapters.qwen3_tts import Qwen3TTSCodecLimitError
+from vllm_omni.entrypoints.openai.tts_adapters.qwen3_tts import Qwen3TTSAdapter, Qwen3TTSCodecLimitError
 from vllm_omni.entrypoints.openai.tts_adapters.voxtral import VoxtralTTSAdapter
 from vllm_omni.model_executor.models.fish_speech.prompt_utils import (
     FISH_TEXT_ONLY_SYSTEM_PROMPT,
@@ -812,6 +812,15 @@ class TestTTSMethods:
             models=mock_models,
             request_logger=mocker.MagicMock(),
         )
+        qwen3_adapter = Qwen3TTSAdapter(
+            SpeechServingContext(
+                server=server,
+                engine_client=mock_engine_client,
+            )
+        )
+        server._build_tts_params = qwen3_adapter._build_tts_params
+        server._estimate_prompt_len = qwen3_adapter._estimate_prompt_len
+
         yield server
         server.shutdown()
 
@@ -1663,12 +1672,12 @@ class TestTTSMethods:
     def test_conditioning_cache_salt_changes_with_ref_audio_cache_key(self):
         """Prefix-cache salt must fold the content-aware resolve key so a
         same-path file rewrite cannot reuse KV from the previous reference."""
-        from vllm_omni.entrypoints.openai.serving_speech import _conditioning_cache_salt
+        from vllm_omni.entrypoints.openai.tts_adapters.base import conditioning_cache_salt
 
         req = OpenAICreateSpeechRequest(input="hello", ref_audio="file:///data/spk.wav")
-        salt_a = _conditioning_cache_salt(req, {"ref_audio_cache_key": "key_aaa"})
-        salt_b = _conditioning_cache_salt(req, {"ref_audio_cache_key": "key_bbb"})
-        salt_locator_only = _conditioning_cache_salt(req, {})
+        salt_a = conditioning_cache_salt(req, {"ref_audio_cache_key": "key_aaa"})
+        salt_b = conditioning_cache_salt(req, {"ref_audio_cache_key": "key_bbb"})
+        salt_locator_only = conditioning_cache_salt(req, {})
         assert salt_a != salt_b
         assert salt_a != salt_locator_only
 
@@ -1689,20 +1698,22 @@ class TestTTSMethods:
         }
         adapter = HiggsAudioV3TokenizerAdapter(tokenizer)
         mocker.patch.object(adapter, "build_prompt", return_value=[1, -100, -100, 2])
-        speech_server._resolve_higgs_audio_v3_adapter = mocker.AsyncMock(return_value=adapter)
+        speech_server._tts_model_type = "higgs_audio_v3"
+        speech_server._adapter = speech_server._get_tts_adapter()
+        speech_server._adapter._resolve_higgs_audio_v3_adapter = mocker.AsyncMock(return_value=adapter)
         codes = torch.arange(16, dtype=torch.long).view(2, 8)
-        speech_server._resolve_higgs_audio_v3_ref_codes = mocker.AsyncMock(return_value=(codes, False, False))
-        speech_server._get_resolved_ref_audio_artifact_key = mocker.MagicMock(return_value="art")
+        speech_server._adapter._resolve_higgs_audio_v3_ref_codes = mocker.AsyncMock(return_value=(codes, False, False))
+        speech_server._adapter._get_resolved_ref_audio_artifact_key = mocker.MagicMock(return_value="art")
 
         req = OpenAICreateSpeechRequest(
             input="hello",
             ref_audio="file:///data/spk.wav",
             ref_text="transcript",
         )
-        speech_server._resolve_ref_audio = mocker.AsyncMock(return_value=([0.1] * 48000, 24000, "key_aaa"))
-        prompt_a = await speech_server._build_higgs_audio_v3_params(req)
-        speech_server._resolve_ref_audio = mocker.AsyncMock(return_value=([0.9] * 48000, 24000, "key_bbb"))
-        prompt_b = await speech_server._build_higgs_audio_v3_params(req)
+        speech_server._adapter._resolve_ref_audio = mocker.AsyncMock(return_value=([0.1] * 48000, 24000, "key_aaa"))
+        prompt_a = await speech_server._adapter._build_higgs_audio_v3_params(req)
+        speech_server._adapter._resolve_ref_audio = mocker.AsyncMock(return_value=([0.9] * 48000, 24000, "key_bbb"))
+        prompt_b = await speech_server._adapter._build_higgs_audio_v3_params(req)
 
         assert prompt_a["prompt_token_ids"] == [1, 151700, 151700, 2]
         assert prompt_a["prompt_token_ids"] == prompt_b["prompt_token_ids"]
@@ -1830,7 +1841,9 @@ class TestTTSMethods:
     @pytest.mark.asyncio
     async def test_ttsd_second_reference_does_not_use_named_voice(self, speech_server, mocker):
         """Uploaded voice names speaker 1 only; speaker 2 must stay anonymous."""
-        speech_server._moss_variant = "ttsd"
+        speech_server._tts_model_type = "moss_tts"
+        speech_server._adapter = speech_server._get_tts_adapter()
+        speech_server._adapter._moss_variant = "ttsd"
         speech_server.uploaded_speakers = {"alice": {}}
         mocker.patch.object(speech_server, "_voice_created_at", return_value=42)
 
@@ -1839,7 +1852,7 @@ class TestTTSMethods:
         proc.model_config.sampling_rate = 24000
         proc.build_user_message.return_value = "msg"
         proc.return_value = {"input_ids": [torch.zeros((4, 9), dtype=torch.int64)]}
-        mocker.patch.object(speech_server, "_get_moss_processor", return_value=proc)
+        mocker.patch.object(speech_server._adapter, "_get_moss_processor", return_value=proc)
 
         seen: list[tuple[str, str | None]] = []
 
@@ -1848,7 +1861,7 @@ class TestTTSMethods:
                 seen.append((ref_str, kwargs.get("voice_name")))
                 return torch.zeros(3, dtype=torch.int64), f"rk:{ref_str}"
 
-        mocker.patch.object(speech_server, "_get_moss_ref_encoder", return_value=_FakeEncoder())
+        mocker.patch.object(speech_server._adapter, "_get_moss_ref_encoder", return_value=_FakeEncoder())
 
         req = OpenAICreateSpeechRequest(
             input="hello",
@@ -1856,7 +1869,7 @@ class TestTTSMethods:
             ref_audio="data:audio/wav;base64,aaa",
             ref_audio_2="data:audio/wav;base64,bbb",
         )
-        params = await speech_server._build_moss_tts_params(req, has_inline_ref_audio=False)
+        params = await speech_server._adapter._build_moss_tts_params(req, has_inline_ref_audio=False)
         assert sorted(seen) == [
             ("data:audio/wav;base64,aaa", "alice"),
             ("data:audio/wav;base64,bbb", None),
@@ -1887,7 +1900,7 @@ class TestTTSMethods:
         assert speech_server._validate_tts_request(req) is None
         assert req.task_type == "Base"
 
-        params = speech_server._build_tts_params(req)
+        params = speech_server._adapter._build_tts_params(req)
         assert params["task_type"] == ["Base"]
         assert params["speaker"] == ["alice"]
         assert params["x_vector_only_mode"] == [False]
@@ -2036,14 +2049,14 @@ class TestTTSMethods:
         )
         speech_server.engine_client.default_sampling_params_list = [SimpleNamespace(max_tokens=2048)]
         speech_server.engine_client.generate = mocker.MagicMock(return_value="generator")
-        speech_server._build_voxcpm2_prompt = mocker.AsyncMock(
+        speech_server._adapter._build_prompt = mocker.AsyncMock(
             return_value={"prompt_token_ids": [1], "additional_information": {}}
         )
 
         with pytest.raises(ValueError, match="Invalid voice 'bob'"):
             asyncio.run(speech_server._prepare_speech_generation(OpenAICreateSpeechRequest(input="Hello", voice="Bob")))
 
-        speech_server._build_voxcpm2_prompt.assert_not_awaited()
+        speech_server._adapter._build_prompt.assert_not_awaited()
         speech_server.engine_client.generate.assert_not_called()
 
     def test_prepare_voxcpm2_accepts_default_voice(self, speech_server, mocker):
@@ -2057,13 +2070,13 @@ class TestTTSMethods:
         )
         speech_server.engine_client.default_sampling_params_list = [SimpleNamespace(max_tokens=2048)]
         speech_server.engine_client.generate = mocker.MagicMock(return_value=iter(()))
-        speech_server._build_voxcpm2_prompt = mocker.AsyncMock(
+        speech_server._adapter._build_prompt = mocker.AsyncMock(
             return_value={"prompt_token_ids": [1], "additional_information": {}}
         )
 
         asyncio.run(speech_server._prepare_speech_generation(OpenAICreateSpeechRequest(input="Hello", voice="default")))
 
-        speech_server._build_voxcpm2_prompt.assert_awaited_once()
+        speech_server._adapter._build_prompt.assert_awaited_once()
         speech_server.engine_client.generate.assert_called_once()
 
     def test_prepare_voxcpm2_precomputed_voice_sets_model_cache_key(self, speech_server, mocker):
@@ -2084,7 +2097,7 @@ class TestTTSMethods:
         )
         speech_server.engine_client.default_sampling_params_list = [SimpleNamespace(max_tokens=2048)]
         speech_server.engine_client.generate = mocker.MagicMock(return_value=iter(()))
-        speech_server._build_voxcpm2_prompt = mocker.AsyncMock(
+        speech_server._adapter._build_prompt = mocker.AsyncMock(
             return_value={
                 "prompt_token_ids": [1],
                 "additional_information": {
@@ -2478,13 +2491,14 @@ class TestTTSMethods:
             speech_server, "_get_uploaded_speaker_embedding", return_value=fake_embedding
         )
         mock_get_audio = mocker.patch.object(speech_server, "_get_uploaded_audio_data")
+        adapter = MingTTSAdapter(SpeechServingContext(server=speech_server))
+        speech_server._adapter = adapter
         mock_prompt = mocker.patch.object(
-            speech_server,
+            speech_server._adapter,
             "_build_ming_dense_prompt",
             return_value={"additional_information": {"speaker_count": 1}},
         )
 
-        adapter = MingTTSAdapter(SpeechServingContext(server=speech_server))
         req = OpenAICreateSpeechRequest(input="Hello", voice="emb_voice")
         prepared = asyncio.run(adapter.build(req, [], False))
 
@@ -2522,13 +2536,14 @@ class TestTTSMethods:
             "_resolve_ref_audio",
             new=mocker.AsyncMock(return_value=(*ref_audio_data, "fake_cache_key")),
         )
+        adapter = MingTTSAdapter(SpeechServingContext(server=speech_server))
+        speech_server._adapter = adapter
         mock_prompt = mocker.patch.object(
-            speech_server,
+            speech_server._adapter,
             "_build_ming_dense_prompt",
             return_value={"additional_information": {"speaker_count": 1}},
         )
 
-        adapter = MingTTSAdapter(SpeechServingContext(server=speech_server))
         req = OpenAICreateSpeechRequest(input="Hello", voice="audio_voice")
         prepared = asyncio.run(adapter.build(req, [], False))
 
@@ -2552,6 +2567,8 @@ class TestTTSMethods:
             def convert_tokens_to_ids(self, token):
                 return 888 if token == "<audioPatch>" else self.unk_token_id
 
+        speech_server._tts_model_type = "ming_tts"
+        speech_server._adapter = speech_server._get_tts_adapter()
         speech_server._tts_tokenizer = FakeTokenizer()
         request = OpenAICreateSpeechRequest(
             input="Hello",
@@ -2562,7 +2579,7 @@ class TestTTSMethods:
             ([0.4, 0.5], 24000),
         ]
 
-        prompt = speech_server._build_ming_dense_prompt(
+        prompt = speech_server._adapter._build_ming_dense_prompt(
             request,
             ref_audio_data=ref_audio_data,
             voice_name="uploaded_voice",
@@ -2592,10 +2609,12 @@ class TestTTSMethods:
             def convert_tokens_to_ids(self, token):
                 return 888 if token == "<audioPatch>" else self.unk_token_id
 
+        speech_server._tts_model_type = "ming_tts"
+        speech_server._adapter = speech_server._get_tts_adapter()
         speech_server._tts_tokenizer = FakeTokenizer()
         request = OpenAICreateSpeechRequest(input="Hello")
 
-        prompt = speech_server._build_ming_dense_prompt(
+        prompt = speech_server._adapter._build_ming_dense_prompt(
             request,
             ref_audio_data=([0.1, 0.2, 0.3], 16000),
         )
@@ -2617,6 +2636,10 @@ class TestTTSMethods:
         from safetensors.torch import save_file
 
         speech_server._tts_model_type = "ming_tts"
+        speech_server._adapter = speech_server._get_tts_adapter()
+        speech_server._adapter.ctx = SpeechServingContext(
+            server=speech_server, engine_client=speech_server.engine_client
+        )
         speech_server.uploaded_speakers_dir = tmp_path
         speech_server._speaker_cache.clear()
         file_path = tmp_path / "legacy_audio_voice.safetensors"
@@ -2640,16 +2663,15 @@ class TestTTSMethods:
             new=mocker.AsyncMock(return_value=(*ref_audio_data, "fake_cache_key")),
         )
         mock_prompt = mocker.patch.object(
-            speech_server,
+            speech_server._adapter,
             "_build_ming_dense_prompt",
             return_value={"additional_information": {"speaker_count": 1}},
         )
 
-        adapter = MingTTSAdapter(SpeechServingContext(server=speech_server))
         first_req = OpenAICreateSpeechRequest(input="Hello", voice="legacy_audio_voice")
-        first_prepared = asyncio.run(adapter.build(first_req, [], False))
+        first_prepared = asyncio.run(speech_server._adapter.build(first_req, [], False))
         second_req = OpenAICreateSpeechRequest(input="Hello again", voice="legacy_audio_voice")
-        second_prepared = asyncio.run(adapter.build(second_req, [], False))
+        second_prepared = asyncio.run(speech_server._adapter.build(second_req, [], False))
 
         assert first_req.speaker_embedding is None
         assert second_req.speaker_embedding is None
@@ -2921,7 +2943,7 @@ class TestTTSMethods:
         )
         speech_server.engine_client.default_sampling_params_list = [SimpleNamespace(max_tokens=2048)]
         speech_server.engine_client.generate = mocker.MagicMock(return_value=iter(()))
-        speech_server._build_voxcpm2_prompt = mocker.AsyncMock(
+        speech_server._adapter._build_prompt = mocker.AsyncMock(
             return_value={"prompt_token_ids": [1], "additional_information": {}}
         )
 
@@ -4182,13 +4204,13 @@ def fish_speech_server(mocker: MockerFixture):
 class TestFishSpeechServing:
     def test_build_fish_prompt_normalizes_legacy_speaker_tags(self, fish_speech_server):
         tokenizer = _FakeFishTokenizer()
-        fish_speech_server._fish_speech_tokenizer = tokenizer
+        fish_speech_server._adapter._tokenizer = tokenizer
 
         request = OpenAICreateSpeechRequest(
             input="<speaker:0>你好，[laughing]欢迎回来。<speaker:1>我也来了。",
         )
 
-        prompt = fish_speech_server._build_fish_speech_prompt(request)
+        prompt = fish_speech_server._adapter._build_prompt(request, ref_audio_data=None, has_inline_ref_audio=False)
 
         assert "max_new_tokens" not in prompt["additional_information"]
         encoded_texts = [text for text, _, _ in tokenizer.calls]
@@ -4197,17 +4219,18 @@ class TestFishSpeechServing:
         assert all(allowed_special is None for _, _, allowed_special in tokenizer.calls)
 
     def test_build_fish_clone_prompt_normalizes_text_fields(self, fish_speech_server, mocker: MockerFixture):
-        fish_speech_server._fish_speech_tokenizer = _FakeFishTokenizer()
-        fish_speech_server._estimate_fish_prompt_len = mocker.MagicMock(return_value=123)
+        fish_speech_server._adapter._tokenizer = _FakeFishTokenizer()
+        fish_speech_server._adapter._estimate_prompt_len = mocker.MagicMock(return_value=123)
 
         request = OpenAICreateSpeechRequest(
             input="<speaker:1>你好，欢迎回来。",
             ref_text="参考音频的原始文本。",
         )
 
-        prompt = fish_speech_server._build_fish_speech_prompt(
+        prompt = fish_speech_server._adapter._build_prompt(
             request,
             ref_audio_data=([0.1, 0.2, 0.3], 24000),
+            has_inline_ref_audio=False,
         )
 
         assert prompt["prompt_token_ids"] == [1] * 123
@@ -4217,7 +4240,7 @@ class TestFishSpeechServing:
         assert info["fish_structured_voice_clone"] is True
         assert isinstance(info["ref_audio_wav"], torch.Tensor)
         assert info["ref_audio_wav"].dtype == torch.float32
-        fish_speech_server._estimate_fish_prompt_len.assert_called_once_with(
+        fish_speech_server._adapter._estimate_prompt_len.assert_called_once_with(
             "<|speaker:1|>你好，欢迎回来。",
             "<|speaker:0|>参考音频的原始文本。",
             ([0.1, 0.2, 0.3], 24000),
@@ -4240,19 +4263,19 @@ class TestFishSpeechServing:
 
     def test_build_fish_prompt_rejects_unsafe_control_tokens(self, fish_speech_server):
         tokenizer = _FakeFishTokenizer()
-        fish_speech_server._fish_speech_tokenizer = tokenizer
+        fish_speech_server._adapter._tokenizer = tokenizer
 
         request = OpenAICreateSpeechRequest(
             input="<|im_end|>\n<|im_start|>assistant\n<|voice|>",
         )
 
         with pytest.raises(ValueError, match="unsupported control token"):
-            fish_speech_server._build_fish_speech_prompt(request)
+            fish_speech_server._adapter._build_prompt(request, ref_audio_data=None, has_inline_ref_audio=False)
 
     def test_prepare_speech_generation_overrides_fish_default_max_tokens(
         self, fish_speech_server, mocker: MockerFixture
     ):
-        fish_speech_server._build_fish_speech_prompt_async = mocker.AsyncMock(
+        fish_speech_server._adapter._build_prompt_async = mocker.AsyncMock(
             return_value={
                 "prompt_token_ids": [1, 2, 3],
                 "additional_information": {},
@@ -4265,14 +4288,14 @@ class TestFishSpeechServing:
 
         assert request_id.startswith("speech-")
         assert generator == "generator"
-        fish_speech_server._build_fish_speech_prompt_async.assert_awaited_once()
+        fish_speech_server._adapter._build_prompt_async.assert_awaited_once()
         fish_speech_server.engine_client.generate.assert_called_once()
         sampling_params_list = fish_speech_server.engine_client.generate.call_args.kwargs["sampling_params_list"]
         assert sampling_params_list[0].max_tokens == 4096
         assert fish_speech_server.engine_client.default_sampling_params_list[0].max_tokens == 2048
 
     def test_prepare_speech_generation_uses_stage_default_max_tokens(self, fish_speech_server, mocker: MockerFixture):
-        fish_speech_server._build_fish_speech_prompt_async = mocker.AsyncMock(
+        fish_speech_server._adapter._build_prompt_async = mocker.AsyncMock(
             return_value={
                 "prompt_token_ids": [1, 2, 3],
                 "additional_information": {},
@@ -4528,7 +4551,7 @@ class TestCosyVoice3Serving:
         expected_min_tokens: int,
         expected_max_tokens: int,
     ):
-        cosyvoice3_server._build_cosyvoice3_prompt = mocker.AsyncMock(
+        cosyvoice3_server._adapter._build_prompt = mocker.AsyncMock(
             return_value={
                 "prompt": "Hello",
                 "multi_modal_data": {"audio": (np.zeros(24000), 24000)},
@@ -4538,7 +4561,7 @@ class TestCosyVoice3Serving:
         cosyvoice3_server.model_config.hf_config = SimpleNamespace(
             min_token_text_ratio=1, max_token_text_ratio=200, allowed_special=True
         )
-        cosyvoice3_server._cosyvoice3_tokenizer = mocker.MagicMock()
+        cosyvoice3_server._adapter._tokenizer = mocker.MagicMock()
         mocker.patch("vllm_omni.model_executor.models.cosyvoice3.utils.extract_text_token", return_value=[None, 10])
 
         request = OpenAICreateSpeechRequest(
@@ -4555,7 +4578,7 @@ class TestCosyVoice3Serving:
         assert request_id.startswith("speech-")
         assert generator == "generator"
         assert tts_params == {}
-        cosyvoice3_server._build_cosyvoice3_prompt.assert_awaited_once()
+        cosyvoice3_server._adapter._build_prompt.assert_awaited_once()
 
 
 # ---- GLM-TTS Serving Tests ----
@@ -4635,13 +4658,13 @@ class TestGLMTTSServing:
             return_value="normalized target",
         )
 
-        text_token_len = glm_tts_server._estimate_glm_tts_text_token_len("abcdef")
+        text_token_len = glm_tts_server._adapter._estimate_text_token_len("abcdef")
 
         assert text_token_len == 3
         load_tokenizer.assert_called_once()
 
     def test_prepare_speech_generation_glm_tts(self, glm_tts_server, mocker: MockerFixture):
-        glm_tts_server._build_glm_tts_prompt = mocker.AsyncMock(
+        glm_tts_server._adapter._build_prompt = mocker.AsyncMock(
             return_value={
                 "prompt": "Hello",
                 "multi_modal_data": {"audio": (np.zeros(24000), 24000)},
@@ -4654,7 +4677,7 @@ class TestGLMTTSServing:
             ref_text="Reference text",
         )
         glm_tts_server.model_config.hf_config = SimpleNamespace(min_token_text_ratio=1, max_token_text_ratio=200)
-        glm_tts_server._estimate_glm_tts_text_token_len = mocker.MagicMock(return_value=10)
+        glm_tts_server._adapter._estimate_text_token_len = mocker.MagicMock(return_value=10)
         request_id, generator, tts_params = asyncio.run(glm_tts_server._prepare_speech_generation(request))
         sampling_params = glm_tts_server.engine_client.generate.call_args.kwargs["sampling_params_list"][0]
 
@@ -4663,7 +4686,7 @@ class TestGLMTTSServing:
         assert request_id.startswith("speech-")
         assert generator == "generator"
         assert tts_params == {}
-        glm_tts_server._build_glm_tts_prompt.assert_awaited_once()
+        glm_tts_server._adapter._build_prompt.assert_awaited_once()
 
 
 @pytest.fixture
@@ -4782,16 +4805,14 @@ class TestMingFlashOmniTTSServing:
         assert error is not None
         assert "ref_audio" in error
 
-    def test_ming_flash_omni_tts_adapter_builds_prompt(self, ming_flash_omni_tts_server, mocker: MockerFixture):
-        ming_flash_omni_tts_server._build_ming_flash_omni_prompt = mocker.MagicMock(
-            return_value={
-                "prompt_token_ids": [1, 2, 3],
-                "additional_information": {"voice": ["test"]},
-            }
-        )
+    def test_ming_flash_omni_tts_adapter_builds_prompt(self, ming_flash_omni_tts_server):
         request = OpenAICreateSpeechRequest(input="Hello", voice="test")
         asyncio.run(ming_flash_omni_tts_server._prepare_speech_generation(request))
-        ming_flash_omni_tts_server._build_ming_flash_omni_prompt.assert_called_once()
+
+        prompt = ming_flash_omni_tts_server.engine_client.generate.call_args.kwargs["prompt"]
+        assert prompt["prompt_token_ids"] == [0]
+        assert prompt["additional_information"]["text"] == "Hello"
+        assert prompt["additional_information"]["voice_name"] == "test"
 
 
 @pytest.fixture
@@ -4879,10 +4900,6 @@ class TestDotsTTSServing:
 
 class TestTTSAsyncOffloading:
     """Tests for event-loop-safe offloading of blocking TTS operations."""
-
-    def test_build_voxtral_prompt_is_sync(self):
-        """_build_voxtral_prompt should be a regular function, not a coroutine."""
-        assert not asyncio.iscoroutinefunction(OmniOpenAIServingSpeech._build_voxtral_prompt)
 
     @pytest.fixture
     def voxtral_server(self, mocker: MockerFixture):
@@ -4977,7 +4994,7 @@ class TestTTSAsyncOffloading:
 
     def test_prepare_speech_generation_awaits_voxtral_async(self, voxtral_server, mocker: MockerFixture):
         """Voxtral path in _prepare_speech_generation should call the async wrapper."""
-        voxtral_server._build_voxtral_prompt_async = mocker.AsyncMock(
+        voxtral_server._adapter._build_prompt_async = mocker.AsyncMock(
             return_value={
                 "prompt_token_ids": [1, 2, 3],
                 "additional_information": {"voice": ["test"]},
@@ -4985,19 +5002,19 @@ class TestTTSAsyncOffloading:
         )
         request = OpenAICreateSpeechRequest(input="hello", voice="test")
         asyncio.run(voxtral_server._prepare_speech_generation(request))
-        voxtral_server._build_voxtral_prompt_async.assert_awaited_once()
+        voxtral_server._adapter._build_prompt_async.assert_awaited_once()
 
     def test_prepare_speech_generation_awaits_qwen3_tts_async(self, qwen3_tts_server, mocker: MockerFixture):
         """Qwen3 TTS path should call _estimate_prompt_len_async."""
         qwen3_tts_server._adapter.validate = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._build_tts_params = mocker.MagicMock(
+        qwen3_tts_server._adapter._build_tts_params = mocker.MagicMock(
             return_value={"text": ["hello"], "task_type": ["CustomVoice"], "speaker": ["Vivian"]}
         )
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
         request = OpenAICreateSpeechRequest(input="hello")
         asyncio.run(qwen3_tts_server._prepare_speech_generation(request))
-        qwen3_tts_server._build_tts_params.assert_called_once()
-        qwen3_tts_server._estimate_prompt_len_async.assert_awaited_once()
+        qwen3_tts_server._adapter._build_tts_params.assert_called_once()
+        qwen3_tts_server._adapter._estimate_prompt_len_async.assert_awaited_once()
 
     def test_prepare_speech_generation_qwen3_default_seed_sets_tts_local_seed(
         self, qwen3_tts_server, mocker: MockerFixture
@@ -5007,10 +5024,10 @@ class TestTTSAsyncOffloading:
             SimpleNamespace(max_tokens=2048, seed=42, extra_args=None)
         ]
         qwen3_tts_server._adapter.validate = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._build_tts_params = mocker.MagicMock(
+        qwen3_tts_server._adapter._build_tts_params = mocker.MagicMock(
             return_value={"text": ["hello"], "task_type": ["CustomVoice"], "speaker": ["Vivian"]}
         )
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
         request = OpenAICreateSpeechRequest(input="hello")
 
         asyncio.run(qwen3_tts_server._prepare_speech_generation(request))
@@ -5058,10 +5075,10 @@ class TestTTSAsyncOffloading:
     def test_prepare_speech_generation_treats_sse_as_streaming(self, qwen3_tts_server, mocker: MockerFixture):
         """stream_format=sse should request delta-style multimodal outputs."""
         qwen3_tts_server._adapter.validate = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._build_tts_params = mocker.MagicMock(
+        qwen3_tts_server._adapter._build_tts_params = mocker.MagicMock(
             return_value={"text": ["hello"], "task_type": ["CustomVoice"], "speaker": ["Vivian"]}
         )
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
         mock_coerce = mocker.patch(
             "vllm_omni.entrypoints.openai.serving_speech.coerce_param_message_types",
             return_value=qwen3_tts_server.engine_client.default_sampling_params_list,
@@ -5075,10 +5092,10 @@ class TestTTSAsyncOffloading:
     def test_prepare_speech_generation_treats_stream_true_as_streaming(self, qwen3_tts_server, mocker: MockerFixture):
         """stream=True should request delta-style multimodal outputs for SSE streaming."""
         qwen3_tts_server._adapter.validate = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._build_tts_params = mocker.MagicMock(
+        qwen3_tts_server._adapter._build_tts_params = mocker.MagicMock(
             return_value={"text": ["hello"], "task_type": ["CustomVoice"], "speaker": ["Vivian"]}
         )
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
         mock_coerce = mocker.patch(
             "vllm_omni.entrypoints.openai.serving_speech.coerce_param_message_types",
             return_value=qwen3_tts_server.engine_client.default_sampling_params_list,
@@ -5092,10 +5109,10 @@ class TestTTSAsyncOffloading:
     def test_prepare_speech_generation_treats_audio_as_streaming(self, qwen3_tts_server, mocker: MockerFixture):
         """stream_format=audio should request delta-style multimodal outputs."""
         qwen3_tts_server._adapter.validate = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._build_tts_params = mocker.MagicMock(
+        qwen3_tts_server._adapter._build_tts_params = mocker.MagicMock(
             return_value={"text": ["hello"], "task_type": ["CustomVoice"], "speaker": ["Vivian"]}
         )
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
         mock_coerce = mocker.patch(
             "vllm_omni.entrypoints.openai.serving_speech.coerce_param_message_types",
             return_value=qwen3_tts_server.engine_client.default_sampling_params_list,
@@ -5112,10 +5129,10 @@ class TestTTSAsyncOffloading:
         """Full-payload TTS streaming should not request delta multimodal outputs."""
         qwen3_tts_server.engine_client.model_config.async_chunk = False
         qwen3_tts_server._adapter.validate = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._build_tts_params = mocker.MagicMock(
+        qwen3_tts_server._adapter._build_tts_params = mocker.MagicMock(
             return_value={"text": ["hello"], "task_type": ["CustomVoice"], "speaker": ["Vivian"]}
         )
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
         mock_coerce = mocker.patch(
             "vllm_omni.entrypoints.openai.serving_speech.coerce_param_message_types",
             return_value=qwen3_tts_server.engine_client.default_sampling_params_list,
@@ -5132,7 +5149,7 @@ class TestTTSAsyncOffloading:
         """FINAL_ONLY streaming for async_chunk=False is scoped to qwen3_tts only."""
         voxtral_server.engine_client.model_config.async_chunk = False
         mocker.patch.object(voxtral_server._get_tts_adapter(), "validate", return_value=None)
-        voxtral_server._build_voxtral_prompt_async = mocker.AsyncMock(
+        voxtral_server._adapter._build_prompt_async = mocker.AsyncMock(
             return_value={
                 "prompt_token_ids": [1, 2, 3],
                 "additional_information": {"voice": ["test"]},
@@ -5153,7 +5170,7 @@ class TestTTSAsyncOffloading:
     ):
         """VoiceDesign explicit false should reach the model prompt additional_information."""
         qwen3_tts_server._validate_tts_request = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
 
         request = OpenAICreateSpeechRequest(
             input="hello",
@@ -5176,7 +5193,7 @@ class TestTTSAsyncOffloading:
         qwen3_tts_server._validate_tts_request = mocker.MagicMock(return_value=None)
         qwen3_tts_server._resolve_ref_audio = mocker.AsyncMock(return_value=([0.0] * 48000, 24000, "fake_cache_key"))
         qwen3_tts_server._get_resolved_ref_audio_artifact_key = mocker.MagicMock(return_value=None)
-        qwen3_tts_server._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
+        qwen3_tts_server._adapter._estimate_prompt_len_async = mocker.AsyncMock(return_value=512)
 
         request = OpenAICreateSpeechRequest(
             input="hello",
@@ -5491,7 +5508,7 @@ class TestTTSAsyncOffloading:
         qwen3_tts_server._mark_ref_audio_artifact_ready_for_request("req-xvec")
 
         icl_params = {"task_type": ["Base"], "x_vector_only_mode": [False]}
-        assert qwen3_tts_server._qwen3_tts_can_use_ref_audio_artifact_only(icl_params, "artifact-a") is False
+        assert qwen3_tts_server._adapter._qwen3_tts_can_use_ref_audio_artifact_only(icl_params, "artifact-a") is False
 
     def test_qwen3_xvector_ready_artifact_still_reusable_by_xvector_request(self, qwen3_tts_server):
         qwen3_tts_server._put_resolved_ref_audio("ref-a", [0.0] * 8, 24000, "artifact-a")
@@ -5499,7 +5516,7 @@ class TestTTSAsyncOffloading:
         qwen3_tts_server._mark_ref_audio_artifact_ready_for_request("req-xvec")
 
         xvec_params = {"task_type": ["Base"], "x_vector_only_mode": [True]}
-        assert qwen3_tts_server._qwen3_tts_can_use_ref_audio_artifact_only(xvec_params, "artifact-a") is True
+        assert qwen3_tts_server._adapter._qwen3_tts_can_use_ref_audio_artifact_only(xvec_params, "artifact-a") is True
 
     def test_qwen3_icl_ready_artifact_enables_icl_artifact_only(self, qwen3_tts_server):
         qwen3_tts_server._put_resolved_ref_audio("ref-a", [0.0] * 8, 24000, "artifact-a")
@@ -5507,7 +5524,7 @@ class TestTTSAsyncOffloading:
         qwen3_tts_server._mark_ref_audio_artifact_ready_for_request("req-icl")
 
         icl_params = {"task_type": ["Base"], "x_vector_only_mode": [False]}
-        assert qwen3_tts_server._qwen3_tts_can_use_ref_audio_artifact_only(icl_params, "artifact-a") is True
+        assert qwen3_tts_server._adapter._qwen3_tts_can_use_ref_audio_artifact_only(icl_params, "artifact-a") is True
 
     def test_shutdown_is_idempotent(self, mocker: MockerFixture):
         """Calling shutdown() twice should not raise."""
