+++ vllm_omni/entrypoints/openai/serving_speech.py
@@ -30,11 +30,9 @@ from vllm.entrypoints.openai.engine.protocol import (
     ErrorResponse,
     RequestResponseMetadata,
 )
-from vllm.inputs import tokens_input
 from vllm.logger import init_logger
 from vllm.multimodal.media import MediaConnector
 from vllm.utils import random_uuid
-from vllm.utils.async_utils import make_async
 from vllm.v1.engine.exceptions import EngineDeadError, EngineGenerateError
 
 from vllm_omni.entrypoints.openai.audio_utils_mixin import AudioMixin, StreamingAudioResampler
@@ -63,17 +61,6 @@ from vllm_omni.entrypoints.openai.tts_adapters import (
     tts_entry_stage_archs,
 )
 from vllm_omni.entrypoints.utils import coerce_param_message_types
-from vllm_omni.model_executor.models.fish_speech.prompt_utils import (
-    build_fish_text_only_prompt_ids,
-    estimate_fish_voice_clone_prompt_len_from_normalized,
-    normalize_fish_voice_clone_texts,
-)
-from vllm_omni.model_executor.models.ming_flash_omni.prompt_utils import (
-    DEFAULT_PROMPT as MING_DEFAULT_PROMPT,
-)
-from vllm_omni.model_executor.models.ming_flash_omni.prompt_utils import (
-    create_instruction as ming_create_instruction,
-)
 from vllm_omni.outputs import OmniRequestOutput
 from vllm_omni.utils.speaker_cache import get_speaker_cache
 
@@ -93,12 +80,6 @@ _SPEECH_USAGE_INPUT_AUDIO_TOKENS_HEADER = "X-VLLM-OMNI-INPUT-AUDIO-TOKENS"
 # ``tts_adapters.detect_tts_model_type``. Adding a TTS model must not require an
 # edit to this module.
 #
-# CosyVoice3 talker expects its reference transcript wrapped in the model
-# instruction template; without the delimiter the talker re-speaks the
-# reference (issue #4644). Matches the offline example/test and upstream demo.
-_COSYVOICE3_PROMPT_DELIMITER = "<|endofprompt|>"
-_COSYVOICE3_PROMPT_PREFIX = f"You are a helpful assistant.{_COSYVOICE3_PROMPT_DELIMITER}"
-
 # Audex contract: zero-codec / invalid generations arrive as empty terminal
 # payloads and must fail the request, never serialize as a successful empty
 # WAV. Covers both the TTS ("audex") and TTA ("audex_tta") pipelines.
@@ -109,12 +90,7 @@ _REF_AUDIO_METADATA_FETCH_ATTEMPTS = 3
 _REMOTE_REF_AUDIO_SCHEMES = frozenset({"http", "https", "data"})
 _REF_AUDIO_RESOLVE_CACHE_MAX_ENTRIES = 256
 _REF_AUDIO_RESOLVE_CACHE_MAX_BYTES = 256 * 1024 * 1024
-_HIGGS_V3_REF_CODE_CACHE_MAX_ENTRIES = 256
-_HIGGS_V3_REF_CODE_CACHE_MAX_BYTES = 64 * 1024 * 1024
-_QWEN3_TTS_REF_AUDIO_CACHE_KEY = "_qwen3_tts_ref_audio_cache_key"
 _TTS_MAX_INSTRUCTIONS_LENGTH = 500
-_TTS_MAX_NEW_TOKENS_MAX = 4096
-_MING_DEFAULT_PROMPT = MING_DEFAULT_PROMPT
 _DEFAULT_VOICE_NAME = "default"
 
 
@@ -220,64 +196,6 @@ def _validate_path_within_directory(file_path: Path, directory: Path) -> bool:
         return False
 
 
-def _conditioning_cache_salt(request, tts_params: dict | None = None) -> str:
-    """Stable hash of the real Stage 0 conditioning for the prefix cache.
-
-    The talker's vLLM prompt is placeholder token ids; the real inputs are
-    rebuilt from text / ref_audio / ref_text into inputs_embeds. vLLM hashes
-    token ids (folded with cache_salt) for prefix caching, so without a salt
-    every request collides and a hit could reuse KV from a semantically
-    different input. Tying the salt to the conditioning keeps a hit safe:
-    identical conditioning may share the prefix, any difference never does.
-
-    Raw request fields alone are not enough for uploaded voices: the request
-    only carries the voice *name* (ref_audio/ref_text/task_type are resolved
-    from stored voice data into ``tts_params``). Delete + re-upload under the
-    same name leaves every raw field identical, so the resolved conditioning
-    must also be folded in. ``voice_created_at`` bumps on every (re-)upload,
-    which uniquely identifies the resolved reference artifact together with
-    the voice name; the decoded ref_audio array itself need not be hashed.
-
-    Local ``ref_audio`` locators are the same class of gap: the raw request
-    hashes the path/URI string, so an on-disk rewrite is invisible to the
-    salt. Callers stash the content-aware key from ``_resolve_ref_audio`` as
-    ``ref_audio_cache_key`` (and ``ref_audio_2_cache_key`` / emotion keys
-    when those clips exist) so a file edit cannot reuse KV from the previous
-    reference.
-    """
-    h = hashlib.sha256()
-    for part in (
-        request.input,
-        request.task_type,
-        request.language,
-        request.voice,
-        request.ref_text,
-        request.ref_audio,
-        request.instructions,
-        request.x_vector_only_mode,
-        request.speaker_embedding,
-    ):
-        h.update(b"\x00")
-        if part is not None:
-            h.update(repr(part).encode("utf-8"))
-    # Fold resolved conditioning that is auto-derived for uploaded voices and
-    # absent from the raw request (or that a locator string cannot see).
-    for key in (
-        "voice_created_at",
-        "task_type",
-        "speaker",
-        "ref_text",
-        "x_vector_only_mode",
-        "ref_audio_cache_key",
-        "ref_audio_2_cache_key",
-    ):
-        h.update(b"\x00")
-        value = tts_params.get(key) if tts_params is not None else None
-        if value is not None:
-            h.update(repr(value).encode("utf-8"))
-    return h.hexdigest()[:32]
-
-
 class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
     _diffusion_mode: bool = False
     _media_connector: MediaConnector | None = None
@@ -306,9 +224,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         # must not satisfy a later ICL request that needs ref_code (#5049).
         self._ref_audio_model_artifact_ready: set[tuple[str, bool]] = set()
         self._request_ref_audio_artifact_keys: dict[str, tuple[str, bool]] = {}
-        self._higgs_audio_v3_ref_code_cache: OrderedDict[str, tuple[torch.Tensor, int]] = OrderedDict()
-        self._higgs_audio_v3_ref_code_cache_bytes = 0
-        self._higgs_audio_v3_ref_code_inflight: dict[str, asyncio.Task[torch.Tensor]] = {}
         self._speaker_cache = get_speaker_cache()
         self._last_upload_ts = 0
         self._upload_lock = asyncio.Lock()
@@ -438,16 +353,14 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         # Find and cache the TTS stage (if any) during initialization
         self._tts_stage = self._find_tts_stage()
         self._is_tts = self._tts_stage is not None
-        self._fish_speech_tokenizer = None
-        self._covo_audio_tokenizer = None
-        # Cached per process: the CosyVoice3 Qwen tokenizer + resolved model
-        # path used for dynamic-token sizing. Without this, every request
-        # re-ran snapshot_download + reloaded the tokenizer (~100 ms on the
-        # TTFP critical path) in the CosyVoice3 adapter's sampling override.
-        self._cosyvoice3_tokenizer = None
 
         # Determine TTS model type or None
         self._tts_model_type = self._detect_tts_model_type()
+
+        # Shared executor for blocking adapter preprocessing. It must exist
+        # before adapter construction so adapters can create their make_async
+        # wrappers during their own lifecycle initialization.
+        self._tts_executor = ThreadPoolExecutor(max_workers=1)
         # Resolve the per-model serving adapter (RFC #4327), keyed on the
         # detected model-type. Every dedicated TTS model has an adapter; the
         # adapter owns request validation, prompt/param building, capability
@@ -467,36 +380,13 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         available_speakers = self._get_available_speakers()
         logger.info("Loaded %d supported speakers: %s", len(available_speakers), sorted(available_speakers))
 
-        # Sub-variant inside the full MOSS-TTS family. We collapse all five
-        # variants onto the same _tts_model_type="moss_tts" because they share
-        # the same stage layout, but request validation + param building
-        # differ per HF repo (voice-clone vs dialogue vs ambient-sound vs
-        # instruction vs streaming voice-clone).
-        self._moss_variant = self._detect_moss_variant() if self._tts_model_type == "moss_tts" else None
-
-        # GLM-TTS lazy-cached resources (populated on first GLM-TTS request)
-        self._glm_tts_text_tokenizer: object | None = None
-        self._glm_tts_text_frontend: object | None = None
-
         # Cache TTS configuration values (computed once, reused per request)
         self._max_instructions_length = self._compute_max_instructions_length()
 
         self._tts_tokenizer = None
-        self._voxcpm2_tokenizer = None
-        self._audex_tokenizer = None
-        self._audex_tta_rvq = None
-        self._voxcpm2_split_map: dict[int, list[int]] = {}
 
         # Batch configuration
         self._batch_max_items: int = getattr(self.engine_client, "tts_batch_max_items", 32)
-
-        # Shared thread pool executor for blocking TTS preprocessing
-        # operations. max_workers=1 serializes tokenizer access to avoid
-        # Rust RefCell "Already borrowed" errors from concurrent use.
-        self._tts_executor = ThreadPoolExecutor(max_workers=1)
-        self._build_voxtral_prompt_async = make_async(self._build_voxtral_prompt, executor=self._tts_executor)
-        self._build_fish_speech_prompt_async = make_async(self._build_fish_speech_prompt, executor=self._tts_executor)
-        self._estimate_prompt_len_async = make_async(self._estimate_prompt_len, executor=self._tts_executor)
 
     def _get_tts_adapter(self):
         """Return the per-model serving adapter for the current ``_tts_model_type``.
@@ -622,73 +512,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         """Get all voice names accepted by the API, including the placeholder default."""
         return self._get_available_speakers() | {_DEFAULT_VOICE_NAME}
 
-    def _estimate_ref_code_len(self, ref_audio: object) -> int | None:
-        """Estimate ref_code length from ref_audio waveform without running the codec.
-
-        The codec produces one frame per (output_sample_rate / encode_downsample_rate)
-        audio samples, so ref_code_len = ceil(duration_seconds * codec_frame_rate).
-        """
-        codec_frame_rate = self._adapter.capabilities.codec_frame_rate
-        if codec_frame_rate is None:
-            return None
-        try:
-            # ref_audio comes from tts_params as [[wav_array, sr]] or similar nested structure
-            item = ref_audio
-            while isinstance(item, list) and item:
-                if len(item) == 2 and isinstance(item[1], (int, float)):
-                    break
-                item = item[0]
-            if isinstance(item, list) and len(item) == 2:
-                wav, sr = item
-            elif isinstance(item, tuple) and len(item) == 2:
-                wav, sr = item
-            else:
-                return None
-            sr = int(sr)
-            if hasattr(wav, "__len__"):
-                n_samples = len(wav)
-            elif hasattr(wav, "shape"):
-                n_samples = wav.shape[-1] if wav.ndim > 1 else wav.shape[0]
-            else:
-                return None
-            if sr <= 0 or n_samples <= 0:
-                return None
-            duration = n_samples / sr
-            return math.ceil(duration * codec_frame_rate)
-        except Exception:
-            return None
-
-    def _estimate_prompt_len(self, tts_params: dict[str, Any]) -> int:
-        """Estimate prompt length so the placeholder matches model-side embeddings."""
-        try:
-            from vllm_omni.model_executor.models.qwen3_tts.prompt_embeds_builder import (
-                Qwen3TTSPromptEmbedsBuilder,
-            )
-
-            if self._tts_tokenizer is None:
-                from transformers import AutoTokenizer
-
-                model_name = self.engine_client.model_config.model
-                self._tts_tokenizer = AutoTokenizer.from_pretrained(
-                    model_name,
-                    trust_remote_code=True,
-                    padding_side="left",
-                )
-            hf_config = self.engine_client.model_config.hf_config
-            talker_config = hf_config.talker_config
-            task_type = (tts_params.get("task_type") or ["CustomVoice"])[0]
-            return Qwen3TTSPromptEmbedsBuilder.estimate_prompt_len_from_additional_information(
-                additional_information=tts_params,
-                task_type=task_type,
-                tokenize_prompt=lambda t: self._tts_tokenizer(t, padding=False)["input_ids"],
-                codec_language_id=getattr(talker_config, "codec_language_id", None),
-                spk_is_dialect=getattr(talker_config, "spk_is_dialect", None),
-                estimate_ref_code_len=self._estimate_ref_code_len,
-            )
-        except Exception as e:
-            logger.warning("Failed to estimate TTS prompt length, using fallback 2048: %s", e)
-            return 2048
-
     def _get_usage_text_tokenizer(self):
         """Return a text tokenizer for counting input-text usage tokens.
 
@@ -784,41 +607,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
             _SPEECH_USAGE_INPUT_AUDIO_TOKENS_HEADER: str(usage.input_token_details.audio_tokens),
         }
 
-    def _estimate_fish_ref_code_len(self, ref_audio: object) -> int | None:
-        """Estimate Fish Speech semantic token length from raw reference audio."""
-        from vllm_omni.model_executor.models.fish_speech.dac_utils import (
-            DAC_HOP_LENGTH,
-            DAC_SAMPLE_RATE,
-        )
-
-        if not isinstance(ref_audio, (list, tuple)) or len(ref_audio) != 2:
-            return None
-        wav, sr = ref_audio
-        sr = int(sr)
-        n_samples = len(wav)
-        if sr <= 0 or n_samples <= 0:
-            return None
-        resampled_len = max(1, math.ceil(n_samples * DAC_SAMPLE_RATE / sr))
-        return max(1, math.ceil(resampled_len / DAC_HOP_LENGTH))
-
-    def _estimate_fish_prompt_len(self, text: str, ref_text: str, ref_audio: object) -> int:
-        """Estimate Fish Speech clone prompt length without encoding reference audio."""
-        try:
-            from transformers import AutoTokenizer
-
-            if self._fish_speech_tokenizer is None:
-                model_name = self.engine_client.model_config.model
-                self._fish_speech_tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-
-            tokenizer = self._fish_speech_tokenizer
-            semantic_len = self._estimate_fish_ref_code_len(ref_audio)
-            if semantic_len is None:
-                raise ValueError("Failed to estimate Fish Speech semantic token length")
-            return estimate_fish_voice_clone_prompt_len_from_normalized(tokenizer, text, ref_text, semantic_len)
-        except Exception as e:
-            logger.warning("Failed to estimate Fish Speech prompt length, using fallback 2048: %s", e)
-            return 2048
-
     def _voice_created_at(self, voice_lower: str) -> int:
         """Return the upload timestamp of an uploaded voice, or 0 for built-ins.
 
@@ -827,41 +615,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         """
         info = self.uploaded_speakers.get(voice_lower)
         return int(info.get("created_at", 0)) if info else 0
-
-    async def _build_voxcpm2_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-        *,
-        uploaded_ref: tuple[np.ndarray, int] | None = None,
-    ) -> dict[str, Any]:
-        """Build prefill prompt for VoxCPM2 TTS (`prompt_token_ids` padded to full prefill length).
-
-        ``uploaded_ref`` supplies the audio for uploaded voices (no explicit
-        ``ref_audio`` in the request) so prefill length includes it.
-        """
-        from vllm_omni.model_executor.models.voxcpm2.voxcpm2_talker import build_voxcpm2_prompt
-
-        self._voxcpm2_encode("")  # lazy-init tokenizer + split_map
-        ref_audio = None
-        ref_sr = None
-        voice_profile = None
-        if request.ref_audio is not None:
-            ref_audio, ref_sr, _ = await self._resolve_ref_audio(request.ref_audio)
-        elif uploaded_ref is not None:
-            wav_np, ref_sr = uploaded_ref
-            ref_audio = wav_np.tolist()
-        elif request.voice is not None:
-            voice_profile = self._adapter.capabilities.precomputed_speakers.get(request.voice.lower())
-        return build_voxcpm2_prompt(
-            hf_config=self.engine_client.model_config.hf_config,
-            tokenizer=self._voxcpm2_tokenizer,
-            split_map=self._voxcpm2_split_map,
-            text=request.input,
-            ref_audio=ref_audio,
-            ref_sr=ref_sr,
-            ref_text=request.ref_text,
-            voice_profile=voice_profile,
-        )
 
     def _load_uploaded_audio(self, voice_name: str) -> tuple[np.ndarray, int] | None:
         """Load decoded audio samples + sample rate from an uploaded voice's safetensors."""
@@ -1330,25 +1083,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
             return "sample_rate is not supported by the current TTS model"
         return None
 
-    def _voxcpm2_encode(self, text: str) -> list[int]:
-        """Tokenize text for VoxCPM2, splitting multichar Chinese tokens."""
-        from vllm_omni.model_executor.models.voxcpm2.voxcpm2_talker import (
-            build_cjk_split_map,
-            split_multichar_chinese,
-        )
-
-        if self._voxcpm2_tokenizer is None:
-            from transformers import AutoTokenizer
-
-            model_name = self.engine_client.model_config.model
-            tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-            self._voxcpm2_split_map = build_cjk_split_map(tokenizer)
-            self._voxcpm2_tokenizer = tokenizer
-            logger.info("VoxCPM2 serving: built multichar split map (%d entries)", len(self._voxcpm2_split_map))
-
-        ids = self._voxcpm2_tokenizer.encode(text, add_special_tokens=True)
-        return split_multichar_chinese(ids, self._voxcpm2_split_map)
-
     def _validate_ref_audio_format(self, ref_audio: str) -> str | None:
         """Validate ref_audio is a supported URI format. Returns error or None."""
         if not isinstance(ref_audio, str):
@@ -1357,445 +1091,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         if scheme not in {"http", "https", "data", "file"}:
             return "ref_audio must be a URL (http/https), base64 data URL (data:...), or file URI (file://...)"
         return None
-
-    def _detect_moss_variant(self) -> str:
-        """Sub-classify a ``moss_tts``-stage server into the actual MOSS-TTS
-        variant family (tts, ttsd, sound_effect, voice_generator, realtime).
-
-        Detection key is the HF repo path / model_name; matches
-        ``_try_resolve_omni_model_type`` in entrypoints/utils.py so users get
-        consistent behaviour no matter how they launched the server (--model
-        OpenMOSS-Team/MOSS-TTSD-v1.0 vs --deploy-config moss_ttsd.yaml).
-        """
-        try:
-            name = (self.engine_client.model_config.model or "").lower().replace("-", "").replace("_", "")
-        except Exception:
-            name = ""
-        if "realtime" in name:
-            return "realtime"
-        if "local" in name:
-            return "local"
-        if "ttsd" in name:
-            return "ttsd"
-        if "soundeffect" in name:
-            return "sound_effect"
-        if "voicegenerator" in name:
-            return "voice_generator"
-        return "tts"
-
-    def _get_moss_processor(self):
-        """Lazily load the upstream MOSS-TTS processor once per server.
-
-        Cached on ``self._moss_processor_cache``. The processor owns its own
-        audio_tokenizer (~1.6 B params); we keep it on CPU so it doesn't
-        compete with the talker (~8 GiB) and codec (~7 GiB) for our 96 GiB
-        GPU — per-request ref-audio encoding is fast enough on CPU.
-        """
-        cached = getattr(self, "_moss_processor_cache", None)
-        if cached is not None:
-            return cached
-        from transformers import AutoProcessor
-
-        model_id = self.engine_client.model_config.model
-        proc = AutoProcessor.from_pretrained(model_id, trust_remote_code=True)
-        if hasattr(proc, "audio_tokenizer"):
-            proc.audio_tokenizer = proc.audio_tokenizer.to("cpu").eval()
-        self._moss_processor_cache = proc
-        return proc
-
-    def _get_moss_ref_encoder(self):
-        """Lazily build the per-server MOSS reference-audio encoder once."""
-        cached = getattr(self, "_moss_ref_encoder", None)
-        if cached is not None:
-            return cached
-        from vllm_omni.model_executor.models.moss_tts.reference_encoder import build_reference_encoder
-
-        # The variant's encode geometry (n_vq, working sample rate) is derived
-        # inside the model package; this layer only supplies the processor and
-        # the process-wide speaker cache.
-        encoder = build_reference_encoder(
-            self._get_moss_processor(),
-            variant=self._moss_variant,
-            speaker_cache=self._speaker_cache,
-        )
-        self._moss_ref_encoder = encoder
-        return encoder
-
-    async def _build_moss_tts_params(
-        self,
-        request: OpenAICreateSpeechRequest,
-        *,
-        has_inline_ref_audio: bool = False,
-    ) -> dict[str, Any]:
-        """Build the talker prompt + ``additional_information`` payload for any
-        MOSS-TTS-family request (nano + 5 full variants).
-
-        For the legacy ``moss_tts_nano`` model_type, keeps the original nano
-        contract (``{text, mode=voice_clone, prompt_audio_array}``); the
-        caller still uses a ``[1]`` placeholder prompt.
-
-        For the full MOSS-TTS family (``MossTTSDelay*`` / ``MossTTSRealtime*``)
-        we **call the upstream processor server-side** to produce the unified
-        ``(text_ids, audio_codes)`` shape the talker actually consumes — same
-        flow as ``examples/.../moss_tts/end2end.py:_build_unified_codes``.
-        Returns ``{prompt_token_ids: list[int], codes.ref: torch.LongTensor,
-        max_new_frames, ...}``. The caller treats ``prompt_token_ids`` as the
-        prompt and forwards the rest as ``additional_information``.
-        """
-        import torch  # local to avoid pulling torch at module import time
-
-        v = self._moss_variant
-
-        # ---- Legacy nano path (unchanged) ----
-        if v is None:  # moss_tts_nano
-            params: dict[str, Any] = {
-                "text": [request.input or ""],
-                "mode": ["voice_clone"],
-            }
-            if request.max_new_tokens is not None:
-                params["max_new_frames"] = [request.max_new_tokens]
-            wav_list, sr, cache_key = await self._resolve_ref_audio(request.ref_audio)
-            params["prompt_audio_array"] = [[wav_list, sr]]
-            params["ref_audio_cache_key"] = cache_key
-            return params
-
-        # ---- MOSS-TTS-Realtime: keep the old prompt_audio_array path ----
-        # ``AutoProcessor.from_pretrained`` doesn't auto-discover
-        # ``MossTTSRealtimeProcessor`` (no ``processor_config.json`` in the
-        # snapshot), and Realtime's prompt format diverges from MossTTSDelay
-        # (16-channel grid, separate per-step text feed). The
-        # ``prompt_audio_array`` shape lines up well enough with what the
-        # talker reads for short prompts; full Realtime support needs a
-        # separate processor.from_module path which we don't wire here.
-        if v == "realtime":
-            params: dict[str, Any] = {
-                "text": [request.input or ""],
-                "mode": ["voice_clone"],
-            }
-            if request.max_new_tokens is not None:
-                params["max_new_frames"] = [request.max_new_tokens]
-            wav_list, sr, cache_key = await self._resolve_ref_audio(request.ref_audio)
-            params["prompt_audio_array"] = [[wav_list, sr]]
-            params["ref_audio_cache_key"] = cache_key
-            return params
-
-        # ---- MossTTSDelay family (tts/ttsd/sound_effect/voice_generator)
-        # and MOSS-TTS-Local-Transformer-v1.5: call the upstream processor
-        # server-side to produce unified codes. Local-v1.5 ships its own
-        # AutoProcessor (processor_config.json + processing_moss_tts.py) and
-        # reuses this exact build_user_message/encode_audios_from_wav path in
-        # the offline example (examples/.../moss_tts/end2end.py:
-        # _build_unified_codes) -- it is NOT in the same boat as Realtime
-        # (no processor_config.json there), so it must not fall back to the
-        # prompt_audio_array path above (which the talker's preprocess()
-        # never reads -- info_dict["codes"]["ref"] is the only thing it
-        # consumes, so skipping this path silently drops all voice-clone
-        # conditioning and produces unconditioned/garbage audio online). ----
-        proc = self._get_moss_processor()
-
-        user_kwargs: dict[str, Any] = {"text": request.input or ""}
-        resolve_keys: dict[int, str] = {}
-        if v in ("tts", "local", "ttsd"):
-            user_kwargs["reference"], resolve_keys = await self._encode_moss_references(
-                request,
-                has_inline_ref_audio=has_inline_ref_audio,
-                two_speaker=(v == "ttsd"),
-            )
-        elif v == "sound_effect":
-            user_kwargs["text"] = request.input or ""  # may be empty
-            user_kwargs["ambient_sound"] = request.ambient_sound or ""
-            if request.duration_seconds is not None:
-                user_kwargs["tokens"] = max(1, int(float(request.duration_seconds) * 12.5))
-            elif request.max_new_tokens is not None:
-                user_kwargs["tokens"] = int(request.max_new_tokens)
-        elif v == "voice_generator":
-            user_kwargs["instruction"] = request.instructions or ""
-
-        # Optional language tag for the spoken-text variants. MOSS-TTS-v1.5's
-        # headline improvement is multilingual synthesis when the language is
-        # given (build_user_message(..., language=...)); 1.0 ignores it
-        # gracefully. Sound-effect output is non-verbal, so skip it there.
-        if v in ("tts", "local", "ttsd", "voice_generator") and getattr(request, "language", None):
-            user_kwargs["language"] = request.language
-
-        # Build the unified-codes prompt: (L, 1+n_vq) where col 0 is text/special
-        # tokens and cols 1..n_vq are the delay-pattern audio code grid (mostly
-        # audio_pad_code outside the reference block).
-        user_msg = proc.build_user_message(**user_kwargs)
-        batch = proc(conversations=[[user_msg]], mode="generation")
-        unified = batch["input_ids"][0]  # torch.LongTensor (L, 1+n_vq)
-        text_ids: list[int] = unified[:, 0].tolist()
-        audio_codes: torch.Tensor = unified[:, 1:].contiguous().to(torch.int64)
-
-        params: dict[str, Any] = {
-            "prompt_token_ids": text_ids,
-            "codes": {"ref": audio_codes},
-        }
-        if request.max_new_tokens is not None:
-            params["max_new_frames"] = [request.max_new_tokens]
-        if 0 in resolve_keys:
-            params["ref_audio_cache_key"] = resolve_keys[0]
-        if 1 in resolve_keys:
-            params["ref_audio_2_cache_key"] = resolve_keys[1]
-        return params
-
-    async def _encode_moss_references(
-        self,
-        request: OpenAICreateSpeechRequest,
-        *,
-        has_inline_ref_audio: bool,
-        two_speaker: bool,
-    ) -> tuple[list, dict[int, str]]:
-        """Encode the request's reference clip(s) into MOSS RVQ code tensors.
-
-        Reference encoding + speaker caching lives in the model package
-        (moss_tts.reference_encoder), mirroring Fish Speech / CosyVoice3 /
-        Qwen3-TTS which keep reference handling with the model rather than in
-        this shared serving file. This method only resolves the serving-side
-        inputs the model helper needs: the generic audio resolver/artifact-key
-        lookups and whether the request may use the named-voice cache.
-
-        Returns ``(codes_per_speaker, resolve_keys)`` where ``resolve_keys``
-        maps the reference slot (0 = ref_audio, 1 = ref_audio_2) to its
-        content-aware resolve key, for salting the KV prefix cache.
-        """
-        from vllm_omni.model_executor.models.moss_tts.reference_encoder import encode_request_references
-
-        # Named-voice caching is only valid for uploaded speakers without an
-        # inline ref_audio: ``request.voice`` plus a file/URL would otherwise
-        # key on (name, created_at=0) and skip the content-aware resolve.
-        raw_voice = getattr(request, "voice", None)
-        raw_voice = raw_voice.strip() if isinstance(raw_voice, str) else ""
-        voice_lower = raw_voice.lower()
-        use_named_voice = bool(voice_lower) and voice_lower in self.uploaded_speakers and not has_inline_ref_audio
-        voice = voice_lower if use_named_voice else ""
-        voice_created = self._voice_created_at(voice) if voice else 0
-
-        return await encode_request_references(
-            self._get_moss_ref_encoder(),
-            request.ref_audio,
-            request.ref_audio_2 if two_speaker else None,
-            resolve_ref_audio=self._resolve_ref_audio,
-            get_artifact_key=self._get_resolved_ref_audio_artifact_key,
-            voice_name=voice or None,
-            voice_created_at=voice_created,
-        )
-
-    async def _build_higgs_audio_v2_params(self, request: OpenAICreateSpeechRequest):
-        """Build prompt_token_ids for higgs_audio_v2 via the upstream processor.
-
-        Plain-text path: runs ``build_plain_text_prompt`` and returns the
-        token-only prompt. Voice-clone path (``ref_audio`` + ``ref_text``):
-        resolves the reference clip via ``_resolve_ref_audio``, runs
-        ``build_voice_clone_prompt`` (which encodes the clip through HF's
-        ``HiggsAudioV2TokenizerModel`` loaded from the k2-fsa/OmniVoice
-        ``audio_tokenizer/`` subdirectory), and attaches the encoded
-        ``audio_input_ids`` + ``audio_input_ids_mask`` tensors via
-        ``additional_information`` so the talker substitutes them at the
-        prompt-side audio placeholders.
-        """
-        from vllm_omni.model_executor.models.higgs_audio_v2.higgs_audio_v2_tokenizer import (
-            build_plain_text_prompt,
-            build_voice_clone_prompt,
-            input_ids_to_python_list,
-        )
-
-        processor = await self._resolve_higgs_audio_v2_processor()
-
-        if request.ref_audio is None:
-            inputs = await asyncio.to_thread(build_plain_text_prompt, processor, request.input)
-            prompt_token_ids = input_ids_to_python_list(inputs)
-            return tokens_input(prompt_token_ids=prompt_token_ids)
-
-        wav_list, sr, _ = await self._resolve_ref_audio(request.ref_audio)
-        wav = np.asarray(wav_list, dtype=np.float32)
-        out = await asyncio.to_thread(
-            build_voice_clone_prompt,
-            processor,
-            request.input,
-            wav,
-            int(sr),
-            request.ref_text or "",
-        )
-        prompt = tokens_input(prompt_token_ids=out["prompt_token_ids"])
-        # Pass tensors at the top level of additional_information (NOT list-
-        # wrapped). ``vllm_omni.data_entry_keys.serialize_payload`` routes
-        # bare ``torch.Tensor`` values through ``_serialize_tensor``; a list
-        # containing tensors would fall into the ``list_data`` field which
-        # msgspec cannot serialize and the tensors would be dropped over the
-        # process boundary (silent voice-clone failure).
-        prompt["additional_information"] = {
-            "audio_input_ids": out["audio_input_ids"],
-            "audio_input_ids_mask": out["audio_input_ids_mask"],
-        }
-        return prompt
-
-    async def _resolve_higgs_audio_v2_processor(self):
-        """Lazy-load the AutoProcessor for higgs_audio_v2 (once per serving instance)."""
-        cached = getattr(self, "_higgs_audio_v2_processor", None)
-        if cached is not None:
-            return cached
-
-        from transformers import AutoProcessor
-
-        model_path = None
-        for stage in self.engine_client.stage_configs:
-            model_path = getattr(getattr(stage, "engine_args", None), "model", None)
-            if model_path:
-                break
-        if model_path is None:
-            # Fallback: the orchestrator stores the served model id on the engine
-            # itself (set by AsyncOmniEngine.__init__). Stage-level engine_args
-            # may not surface ``model`` when the deploy yaml doesn't set it per
-            # stage (the CLI-passed model id is the single source of truth).
-            model_path = getattr(self.engine_client, "model", None)
-        if model_path is None:
-            raise RuntimeError("higgs_audio_v2 serving could not resolve the model path from the engine stage configs")
-        processor = AutoProcessor.from_pretrained(model_path)
-        self._higgs_audio_v2_processor = processor
-        return processor
-
-    # ---- higgs-audio v3 ----
-
-    async def _build_higgs_audio_v3_params(self, request: OpenAICreateSpeechRequest):
-        """Build prompt_token_ids for higgs_audio_v3.
-
-        Plain-text path: builds ``[tts, text, tokens, audio]``.
-        Voice-clone path: encodes reference audio, applies delay pattern, and
-        records reference-audio prompt positions while submitting valid token IDs.
-        """
-        adapter = await self._resolve_higgs_audio_v3_adapter()
-
-        if request.ref_audio is None:
-            prompt_ids = adapter.build_prompt(request.input)
-            return tokens_input(prompt_token_ids=prompt_ids)
-
-        # Voice clone
-        from vllm_omni.model_executor.models.higgs_audio_v3.higgs_audio_v3_tokenizer import (
-            apply_delay_pattern,
-            encode_reference_audio,
-        )
-
-        wav_list, sr, cache_key = await self._resolve_ref_audio(request.ref_audio)
-        artifact_key = self._get_resolved_ref_audio_artifact_key(cache_key)
-        wav = np.asarray(wav_list, dtype=np.float32)
-        ref_codes_delayed, cache_hit, inflight_wait = await self._resolve_higgs_audio_v3_ref_codes(
-            artifact_key,
-            wav,
-            int(sr),
-            encode_reference_audio,
-            apply_delay_pattern,
-        )
-        del cache_hit, inflight_wait
-
-        prompt_ids = adapter.build_prompt(
-            request.input,
-            num_ref_tokens=int(ref_codes_delayed.shape[0]),
-            reference_text=request.ref_text or None,
-        )
-        prompt_ids, audio_placeholder_positions = adapter.prepare_prompt_for_engine(prompt_ids)
-        prompt = tokens_input(prompt_token_ids=prompt_ids)
-        import torch
-
-        prompt["additional_information"] = {
-            "audio_input_ids": ref_codes_delayed.to(torch.long),
-            "audio_input_ids_mask": torch.ones(ref_codes_delayed.shape[0], dtype=torch.bool),
-            "audio_placeholder_positions": audio_placeholder_positions,
-            "ref_audio_cache_key": cache_key,
-        }
-        # Placeholder prompt ids do not change when a same-path file is
-        # rewritten at the same duration; fold the content-aware resolve key
-        # into cache_salt so prefix caching cannot reuse the previous KV.
-        prompt["cache_salt"] = _conditioning_cache_salt(request, prompt["additional_information"])
-        return prompt
-
-    async def _resolve_higgs_audio_v3_ref_codes(
-        self,
-        artifact_key: str | None,
-        wav: np.ndarray,
-        sr: int,
-        encode_reference_audio,
-        apply_delay_pattern,
-    ) -> tuple[torch.Tensor, bool, bool]:
-        ref_codes_delayed = self._get_higgs_audio_v3_ref_codes(artifact_key)
-        if ref_codes_delayed is not None:
-            return ref_codes_delayed, True, False
-        if not artifact_key:
-            ref_codes_raw = await asyncio.to_thread(encode_reference_audio, wav, sr)
-            return apply_delay_pattern(ref_codes_raw), False, False
-
-        task = self._higgs_audio_v3_ref_code_inflight.get(artifact_key)
-        if task is not None:
-            return (await task).clone(), False, True
-
-        async def _encode_and_cache() -> torch.Tensor:
-            ref_codes_raw = await asyncio.to_thread(encode_reference_audio, wav, sr)
-            delayed = apply_delay_pattern(ref_codes_raw)
-            self._put_higgs_audio_v3_ref_codes(artifact_key, delayed)
-            cached = self._get_higgs_audio_v3_ref_codes(artifact_key)
-            return cached if cached is not None else delayed.detach().to("cpu", dtype=torch.long).contiguous()
-
-        task = asyncio.create_task(_encode_and_cache())
-        self._higgs_audio_v3_ref_code_inflight[artifact_key] = task
-        try:
-            return (await task).clone(), False, False
-        finally:
-            if self._higgs_audio_v3_ref_code_inflight.get(artifact_key) is task:
-                self._higgs_audio_v3_ref_code_inflight.pop(artifact_key, None)
-
-    def _get_higgs_audio_v3_ref_codes(self, artifact_key: str | None) -> torch.Tensor | None:
-        if not artifact_key:
-            return None
-        cached = self._higgs_audio_v3_ref_code_cache.get(artifact_key)
-        if cached is None:
-            return None
-        self._higgs_audio_v3_ref_code_cache.move_to_end(artifact_key)
-        return cached[0].clone()
-
-    def _put_higgs_audio_v3_ref_codes(self, artifact_key: str, codes: torch.Tensor) -> None:
-        if _HIGGS_V3_REF_CODE_CACHE_MAX_ENTRIES <= 0 or _HIGGS_V3_REF_CODE_CACHE_MAX_BYTES <= 0 or not artifact_key:
-            return
-        cached_codes = codes.detach().to("cpu", dtype=torch.long).contiguous()
-        size = int(cached_codes.numel() * cached_codes.element_size())
-        if size > _HIGGS_V3_REF_CODE_CACHE_MAX_BYTES:
-            return
-        previous = self._higgs_audio_v3_ref_code_cache.pop(artifact_key, None)
-        if previous is not None:
-            self._higgs_audio_v3_ref_code_cache_bytes -= previous[1]
-        self._higgs_audio_v3_ref_code_cache[artifact_key] = (cached_codes, size)
-        self._higgs_audio_v3_ref_code_cache_bytes += size
-        while len(self._higgs_audio_v3_ref_code_cache) > _HIGGS_V3_REF_CODE_CACHE_MAX_ENTRIES:
-            _, (_, old_size) = self._higgs_audio_v3_ref_code_cache.popitem(last=False)
-            self._higgs_audio_v3_ref_code_cache_bytes -= old_size
-        while self._higgs_audio_v3_ref_code_cache_bytes > _HIGGS_V3_REF_CODE_CACHE_MAX_BYTES:
-            _, (_, old_size) = self._higgs_audio_v3_ref_code_cache.popitem(last=False)
-            self._higgs_audio_v3_ref_code_cache_bytes -= old_size
-
-    async def _resolve_higgs_audio_v3_adapter(self):
-        """Lazy-load the tokenizer adapter for higgs_audio_v3."""
-        cached = getattr(self, "_higgs_audio_v3_adapter", None)
-        if cached is not None:
-            return cached
-
-        from transformers import AutoTokenizer
-
-        from vllm_omni.model_executor.models.higgs_audio_v3.higgs_audio_v3_tokenizer import (
-            HiggsAudioV3TokenizerAdapter,
-        )
-
-        model_path = None
-        for stage in self.engine_client.stage_configs:
-            model_path = getattr(getattr(stage, "engine_args", None), "model", None)
-            if model_path:
-                break
-        if model_path is None:
-            model_path = getattr(self.engine_client, "model", None)
-        if model_path is None:
-            raise RuntimeError("higgs_audio_v3 serving could not resolve model path")
-        tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
-        adapter = HiggsAudioV3TokenizerAdapter(tokenizer)
-        self._higgs_audio_v3_adapter = adapter
-        return adapter
 
     @staticmethod
     def _local_ref_audio_stat_path(ref_audio_str: str) -> str | None:
@@ -2043,14 +1338,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
     def _tts_x_vector_only(tts_params: dict[str, Any]) -> bool:
         return bool((tts_params.get("x_vector_only_mode") or [False])[0])
 
-    def _qwen3_tts_can_use_ref_audio_artifact_only(self, tts_params: dict[str, Any], artifact_key: str | None) -> bool:
-        if self._tts_model_type != "qwen3_tts":
-            return False
-        x_vector_only = self._tts_x_vector_only(tts_params)
-        if not artifact_key or (artifact_key, x_vector_only) not in self._ref_audio_model_artifact_ready:
-            return False
-        return (tts_params.get("task_type") or ["CustomVoice"])[0] == "Base"
-
     def _track_ref_audio_artifact_warmup(
         self, request_id: str, artifact_key: str | None, x_vector_only: bool = False
     ) -> None:
@@ -2074,145 +1361,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
             wav_list, sr, _ = await self._resolve_ref_audio(ref_audio)
             resolved.append((wav_list, sr))
         return resolved
-
-    # ---- Ming TTS helpers ----
-
-    def _coerce_ming_prompt_waveform(self, wav_samples, sample_rate):
-        from torchaudio.functional import resample as resample_audio
-
-        from vllm_omni.model_executor.models.ming_tts.config_ming_tts import SAMPLE_RATE
-
-        waveform = torch.as_tensor(wav_samples, dtype=torch.float32).reshape(1, -1)
-        if int(sample_rate) != SAMPLE_RATE:
-            waveform = resample_audio(waveform, int(sample_rate), SAMPLE_RATE)
-        return waveform
-
-    def _build_ming_reference_waveforms(
-        self,
-        ref_audio_data: tuple[list[float], int] | list[tuple[list[float], int]] | None,
-    ) -> tuple[list[torch.Tensor], list[int]]:
-        if ref_audio_data is None:
-            return [], []
-        items = ref_audio_data if isinstance(ref_audio_data, list) else [ref_audio_data]
-        waveforms = [torch.as_tensor(samples, dtype=torch.float32).reshape(1, -1) for samples, _ in items]
-        sample_rates = [int(sample_rate) for _, sample_rate in items]
-        if any(sample_rate <= 0 for sample_rate in sample_rates):
-            raise ValueError(f"Ming reference audio sample rates must be positive, got {sample_rates}")
-        return waveforms, sample_rates
-
-    def _parse_ming_instruction_fields(
-        self,
-        request,
-        *,
-        include_voice=False,
-        plain_text_passthrough=False,
-    ):
-        instruction_text = request.instructions.strip() if isinstance(request.instructions, str) else None
-        instruction_dict: dict[str, Any] = {}
-
-        voice_lower = request.voice.lower() if isinstance(request.voice, str) else None
-        if include_voice and request.voice and not (voice_lower and voice_lower in self.uploaded_speakers):
-            instruction_dict["IP"] = request.voice
-
-        if instruction_text:
-            try:
-                parsed = json.loads(instruction_text)
-            except json.JSONDecodeError:
-                parsed = None
-            if isinstance(parsed, dict):
-                instruction_dict.update(parsed)
-            elif instruction_dict or not plain_text_passthrough:
-                instruction_dict["风格"] = instruction_text
-            else:
-                return instruction_text
-
-        return instruction_dict or None
-
-    def _parse_ming_instruction(self, request: OpenAICreateSpeechRequest) -> Any:
-        """Build a Ming instruction payload from OpenAI speech fields."""
-        return self._parse_ming_instruction_fields(
-            request,
-            include_voice=True,
-            plain_text_passthrough=True,
-        )
-
-    def _build_ming_dense_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-        *,
-        ref_audio_data: tuple[list[float], int] | list[tuple[list[float], int]] | None = None,
-        voice_name: str | None = None,
-        voice_created_at: int = 0,
-    ) -> dict[str, Any]:
-        """Build a Ming dense prompt directly from the OpenAI speech request."""
-        from transformers import AutoTokenizer
-
-        from vllm_omni.model_executor.models.ming_tts.config_ming_tts import (
-            KEY_MAX_DECODE_STEPS,
-            KEY_SPEAKER_SAMPLE_RATES,
-            KEY_SPEAKER_WAVEFORM,
-            KEY_SPEAKER_WAVEFORM_LENGTHS,
-        )
-        from vllm_omni.model_executor.models.ming_tts.prompt_assembly import build_ming_dense_prompt
-
-        if self._tts_tokenizer is None:
-            model_name = self.engine_client.model_config.model
-            trust_remote_code = bool(getattr(self.engine_client.model_config, "trust_remote_code", False))
-            self._tts_tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=trust_remote_code)
-
-        ref_text = request.ref_text
-        reference_waveforms, reference_sample_rates = self._build_ming_reference_waveforms(ref_audio_data)
-        reference_waveform = torch.cat(reference_waveforms, dim=-1) if reference_waveforms else None
-        prompt_waveforms = (
-            [
-                self._coerce_ming_prompt_waveform(waveform, sample_rate)
-                for waveform, sample_rate in zip(reference_waveforms, reference_sample_rates, strict=True)
-            ]
-            if ref_text is not None
-            else []
-        )
-        prompt_waveform = torch.cat(prompt_waveforms, dim=-1) if prompt_waveforms else None
-        speaker_embedding = request.speaker_embedding
-        pending_speaker_extraction = speaker_embedding is None and reference_waveform is not None
-        use_zero_spk_emb = not pending_speaker_extraction and prompt_waveform is None and speaker_embedding is None
-
-        runtime_controls = {}
-        if request.max_new_tokens is not None:
-            runtime_controls[KEY_MAX_DECODE_STEPS] = request.max_new_tokens
-
-        prompt_dict = build_ming_dense_prompt(
-            self._tts_tokenizer,
-            # bgm / music-prompt mode not supported online;
-            # requires prompt_mode API extension (deferred).
-            prompt=_MING_DEFAULT_PROMPT,
-            text=request.input,
-            runtime_controls=runtime_controls or None,
-            instruction=self._parse_ming_instruction(request),
-            prompt_text=ref_text,
-            prompt_waveform=prompt_waveform,
-            speaker_embedding=speaker_embedding,
-            speaker_count=len(reference_waveforms) if pending_speaker_extraction else None,
-            use_zero_spk_emb=use_zero_spk_emb,
-        )
-        additional_information = prompt_dict["additional_information"]
-        if pending_speaker_extraction:
-            additional_information[KEY_SPEAKER_WAVEFORM] = reference_waveform
-            additional_information[KEY_SPEAKER_WAVEFORM_LENGTHS] = torch.tensor(
-                [int(waveform.shape[-1]) for waveform in reference_waveforms],
-                dtype=torch.int32,
-            )
-            additional_information[KEY_SPEAKER_SAMPLE_RATES] = torch.tensor(
-                reference_sample_rates,
-                dtype=torch.int32,
-            )
-            if voice_name:
-                additional_information["voice_name"] = voice_name
-                additional_information["voice_created_at"] = int(voice_created_at)
-        prompt = tokens_input(prompt_token_ids=prompt_dict["prompt_token_ids"])
-        prompt["prompt"] = prompt_dict["prompt"]
-        prompt["text"] = prompt_dict["text"]
-        prompt["additional_information"] = prompt_dict["additional_information"]
-        return prompt
 
     async def _generate_audio_chunks(
         self,
@@ -2542,534 +1690,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
         key = "audio" if "audio" in mm else ("model_outputs" if "model_outputs" in mm else None)
         return mm, key
 
-    def _build_tts_params(self, request: OpenAICreateSpeechRequest) -> dict[str, Any]:
-        """Build TTS parameters from request.
-
-        Processes each parameter if present, skips if not.
-        Values are wrapped in lists as required by the model.
-        """
-        params: dict[str, Any] = {}
-
-        # Text content (always required)
-        params["text"] = [request.input]
-
-        # Task type
-        if request.task_type is not None:
-            params["task_type"] = [request.task_type]
-        else:
-            params["task_type"] = ["CustomVoice"]
-
-        # Language
-        if request.language is not None:
-            params["language"] = [request.language]
-        else:
-            params["language"] = ["Auto"]
-
-        # Speaker (voice)
-        if request.voice is not None:
-            voice_lower = request.voice.lower()
-            precomputed_speakers = self._adapter.capabilities.precomputed_speakers if self._adapter is not None else {}
-            params["speaker"] = [request.voice]
-            params["voice_created_at"] = [self._voice_created_at(voice_lower)]
-
-            # Uploaded voices use task_type="Base" (CustomVoice requires built-in spk_id).
-            # If ref_text was provided at upload time, use in-context cloning; otherwise x_vector only.
-            if voice_lower in self.uploaded_speakers and request.ref_audio is None:
-                speaker_info = self.uploaded_speakers[voice_lower]
-
-                # Check if this voice was uploaded with a pre-computed embedding.
-                # Populate request.speaker_embedding so the existing code path
-                # (below) handles voice_clone_prompt and x_vector_only_mode.
-                embedding = self._get_uploaded_speaker_embedding(request.voice)
-                if embedding is not None:
-                    request.speaker_embedding = embedding
-                    params["speaker"] = [voice_lower]
-                    params["task_type"] = ["Base"]
-                    logger.info("Auto-set speaker_embedding for uploaded voice: %s", request.voice)
-                else:
-                    audio_data = self._get_uploaded_audio_data(request.voice)
-                    if not audio_data:
-                        raise ValueError(f"Audio file for uploaded voice '{request.voice}' is missing or corrupted")
-                    stored_ref_text = speaker_info.get("ref_text")
-                    params["speaker"] = [voice_lower]
-                    params["ref_audio"] = [audio_data]
-                    params["task_type"] = ["Base"]
-                    if stored_ref_text:
-                        params["ref_text"] = [stored_ref_text]
-                        params["x_vector_only_mode"] = [False]
-                    else:
-                        params["x_vector_only_mode"] = [True]
-                    logger.info(
-                        "Auto-set ref_audio for uploaded voice: %s (icl=%s)", request.voice, bool(stored_ref_text)
-                    )
-            elif voice_lower in precomputed_speakers and request.ref_audio is None:
-                profile = precomputed_speakers[voice_lower]
-                mode = str(profile.get("mode") or "xvec").lower()
-                params["speaker"] = [voice_lower]
-                params["task_type"] = ["Base"]
-                params["x_vector_only_mode"] = [mode != "icl"]
-                ref_text = request.ref_text or profile.get("ref_text")
-                if isinstance(ref_text, str) and ref_text.strip():
-                    params["ref_text"] = [ref_text]
-                ref_code_length = profile.get("ref_code_length")
-                if mode == "icl" and ref_code_length:
-                    params["ref_code_length"] = [int(ref_code_length)]
-                logger.info("Using precomputed Qwen3-TTS custom voice profile: %s (mode=%s)", voice_lower, mode)
-
-        elif params["task_type"][0] == "CustomVoice":
-            params["speaker"] = ["Vivian"]  # Default for CustomVoice
-
-        # Instructions for style/emotion control
-        if request.instructions is not None:
-            params["instruct"] = [request.instructions]
-        else:
-            params["instruct"] = [""]
-
-        # Voice clone: ref_audio resolved in create_speech(), not here.
-        if request.ref_text is not None:
-            params["ref_text"] = [request.ref_text]
-        if request.speaker_embedding is not None:
-            # Store as plain float list (not tensor) so it survives msgspec
-            # serialization through the EngineCore IPC boundary.  The talker's
-            # _build_prompt_embeds converts it back to a tensor on the GPU.
-            params["voice_clone_prompt"] = [
-                {
-                    "ref_spk_embedding": list(request.speaker_embedding),
-                }
-            ]
-            # speaker_embedding implies x_vector_only_mode
-            params["x_vector_only_mode"] = [True]
-        elif request.x_vector_only_mode is not None:
-            params["x_vector_only_mode"] = [request.x_vector_only_mode]
-
-        # Generation parameters
-        if request.max_new_tokens is not None:
-            params["max_new_tokens"] = [request.max_new_tokens]
-        else:
-            params["max_new_tokens"] = [2048]
-
-        if request.initial_codec_chunk_frames is not None:
-            params["initial_codec_chunk_frames"] = [request.initial_codec_chunk_frames]
-
-        if request.non_streaming_mode is not None:
-            params["non_streaming_mode"] = [request.non_streaming_mode]
-        # Preserve the legacy VoiceDesign fallback when the request omits an
-        # explicit override. CustomVoice and Base rely on model defaults
-        # (True and False respectively).
-        elif params["task_type"][0] == "VoiceDesign":
-            params["non_streaming_mode"] = [True]
-
-        return params
-
-    # ---- Voxtral TTS helpers ----
-
-    def _build_voxtral_prompt(self, request: OpenAICreateSpeechRequest) -> dict[str, Any]:
-        """Build Voxtral TTS engine prompt, supporting both preset voices and inline
-        ``ref_audio`` (base64 or data URI)."""
-        from mistral_common.protocol.speech.request import SpeechRequest
-
-        text = request.input
-        voice = request.voice
-        ref_audio = request.ref_audio
-        if not voice and not ref_audio:
-            raise ValueError("Voxtral requires either a voice name or ref_audio.")
-        # mistral_common expects raw base64 (no data: prefix)
-        if ref_audio is not None and isinstance(ref_audio, str) and ref_audio.startswith("data:"):
-            _, _, ref_audio = ref_audio.partition(",")
-        if self._tts_tokenizer is None:
-            from vllm.tokenizers import cached_tokenizer_from_config
-
-            mistral_tokenizer = cached_tokenizer_from_config(self.engine_client.model_config)
-            self._tts_tokenizer = mistral_tokenizer.instruct
-        if voice is not None:
-            tokens = self._tts_tokenizer.encode_speech_request(SpeechRequest(input=text, voice=voice)).tokens
-            prompt = tokens_input(prompt_token_ids=tokens)
-            prompt["additional_information"] = {"voice": [voice]}
-            return prompt
-        else:
-            tokenized = self._tts_tokenizer.encode_speech_request(SpeechRequest(input=text, ref_audio=ref_audio))
-            audio = tokenized.audios[0]
-            return {
-                "prompt_token_ids": tokenized.tokens,
-                "multi_modal_data": {"audio": [(audio.audio_array, audio.sampling_rate)]},
-            }
-
-    # ---- Step-Audio2 helpers ----
-
-    def _build_step_audio2_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-    ) -> dict[str, Any]:
-        """Build prompt for Step-Audio2 TTS.
-
-        Constructs the chat prompt with ``<tts_start>`` as the last token
-        of the assistant turn (without ``<|im_end|>``), so the thinker
-        continues generating audio tokens.
-
-        Prompt format::
-            <|im_start|>system\\n{system_prompt}<|im_end|>\\n
-            <|im_start|>user\\n{input_text}<|im_end|>\\n
-            <|im_start|>assistant\\n<tts_start>
-        """
-        system_prompt = getattr(request, "instructions", None) or "You are a voice assistant. Read the text aloud."
-        text = request.input
-
-        raw_prompt = (
-            f"<|im_start|>system\n{system_prompt}<|im_end|>\n"
-            f"<|im_start|>user\n{text}<|im_end|>\n"
-            f"<|im_start|>assistant\n<tts_start>"
-        )
-        return {"prompt": raw_prompt}
-
-    # ---- Fish Speech helpers ----
-
-    def _build_fish_speech_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-        ref_audio_data: tuple[list[float], int] | None = None,
-        *,
-        has_inline_ref_audio: bool = False,
-    ) -> dict[str, Any]:
-        """Build prompt for Fish Speech S2 Pro.
-
-        Without voice cloning:
-          <|im_start|>system\\nconvert the provided text to speech<|im_end|>
-          <|im_start|>user\\n{text}<|im_end|>\\n<|im_start|>assistant\\n<|voice|>
-
-        With voice cloning (ref_audio + ref_text):
-          <|im_start|>system\\nconvert the provided text to speech reference to the following...
-          <|im_end|>\\n<|im_start|>user\\n{text}<|im_end|>\\n<|im_start|>assistant\\n<|voice|>
-        """
-        from transformers import AutoTokenizer
-
-        if self._fish_speech_tokenizer is None:
-            model_name = self.engine_client.model_config.model
-            self._fish_speech_tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-
-        tokenizer = self._fish_speech_tokenizer
-
-        if ref_audio_data is None or not request.ref_text:
-            prompt_ids, normalized_text = build_fish_text_only_prompt_ids(tokenizer, request.input)
-
-            # Keep the prompt-dict metadata shape aligned with the existing text-only
-            # TTS entrypoints: scalar values are wrapped in single-item lists before
-            # EngineCore serialization. Structured clone below is different because
-            # model-side preprocess consumes concrete per-request scalar fields.
-            additional_information: dict[str, Any] = {
-                "text": [normalized_text],
-            }
-            if request.max_new_tokens is not None:
-                additional_information["max_new_tokens"] = [request.max_new_tokens]
-            prompt = tokens_input(prompt_token_ids=prompt_ids)
-            prompt["additional_information"] = additional_information
-            return prompt
-
-        wav_samples, sr = ref_audio_data
-        normalized_text, normalized_ref_text = normalize_fish_voice_clone_texts(request.input, request.ref_text)
-        ph_len = self._estimate_fish_prompt_len(normalized_text, normalized_ref_text, ref_audio_data)
-
-        # Structured clone: scalars (not list-wrapped) because model-side
-        # preprocess() consumes per-request fields directly.
-        additional_information: dict[str, Any] = {
-            "text": normalized_text,
-            "ref_text": normalized_ref_text,
-            "ref_audio_wav": torch.from_numpy(np.asarray(wav_samples, dtype=np.float32)),
-            "ref_audio_sr": int(sr),
-            "fish_structured_voice_clone": True,
-        }
-        # Pass voice identity for model-side DAC code caching.
-        if request.voice is not None:
-            voice_lower = request.voice.lower()
-            if voice_lower in self.uploaded_speakers and not has_inline_ref_audio:
-                additional_information["voice_name"] = voice_lower
-                additional_information["voice_created_at"] = self._voice_created_at(voice_lower)
-        if request.max_new_tokens is not None:
-            additional_information["max_new_tokens"] = request.max_new_tokens
-        prompt = tokens_input(prompt_token_ids=[1] * ph_len)
-        prompt["additional_information"] = additional_information
-        return prompt
-
-    # ---- CosyVoice3 helpers ----
-
-    async def _build_cosyvoice3_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-        *,
-        has_inline_ref_audio: bool = False,
-    ) -> dict[str, Any]:
-        """Build prompt for CosyVoice3.
-
-        CosyVoice3 uses multimodal input with reference audio for voice cloning.
-        The prompt format matches the offline example: text prompt + audio data
-        + mm_processor_kwargs with prompt_text.
-        """
-        # Resolve reference audio
-        wav_samples, sr, _ = await self._resolve_ref_audio(request.ref_audio)
-        audio_data = (np.asarray(wav_samples, dtype=np.float32), sr)
-
-        # Wrap the reference transcript in the CosyVoice3 instruction template
-        # so the talker emits target-only speech (see _COSYVOICE3_PROMPT_PREFIX).
-        # Skip if the caller already supplied a formatted prompt_text.
-        ref_text = request.ref_text or ""
-        if _COSYVOICE3_PROMPT_DELIMITER not in ref_text:
-            ref_text = f"{_COSYVOICE3_PROMPT_PREFIX}{ref_text}"
-        mm_kwargs: dict[str, Any] = {
-            "prompt_text": ref_text,
-            "sample_rate": sr,
-        }
-        # Pass voice metadata for caching in the processor
-        if request.voice:
-            voice_lower = request.voice.lower()
-            if voice_lower in self.uploaded_speakers and not has_inline_ref_audio:
-                mm_kwargs["voice_name"] = voice_lower
-                mm_kwargs["voice_created_at"] = self._voice_created_at(voice_lower)
-
-        return {
-            "prompt": request.input,
-            "multi_modal_data": {
-                "audio": audio_data,
-            },
-            "mm_processor_kwargs": mm_kwargs,
-        }
-
-    # ---- Covo-Audio helpers ----
-
-    def _build_covo_audio_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-    ) -> dict[str, Any]:
-        """Build a chat-style prompt for Covo-Audio-Chat.
-
-        Covo-Audio requires a specific system prompt that instructs the model
-        to interleave text and audio tokens in its output.  We render the
-        messages through the chat template and pass prompt_token_ids so that
-        the engine does not need to re-tokenize.
-        """
-        from transformers import AutoTokenizer
-
-        from vllm_omni.model_executor.models.covo_audio.prompt_utils import (
-            build_covo_audio_prompt_token_ids,
-        )
-
-        if self._covo_audio_tokenizer is None:
-            model_name = self.engine_client.model_config.model
-            try:
-                self._covo_audio_tokenizer = AutoTokenizer.from_pretrained(
-                    model_name,
-                    trust_remote_code=True,
-                )
-            except Exception as exc:
-                raise RuntimeError(f"Failed to load Covo-Audio tokenizer from '{model_name}': {exc}") from exc
-
-        prompt_ids = build_covo_audio_prompt_token_ids(
-            self._covo_audio_tokenizer,
-            request.input,
-        )
-        return {"prompt_token_ids": prompt_ids}
-
-    def _get_audex_tokenizer(self):
-        """Load (once per process) the Audex thinker tokenizer for CFG/TTA prompts.
-
-        The folder follows the serving stage: the full-pipeline thinker
-        (``audex_omni``) tokenizes with ``checkpoint_folder_full``; the TTS
-        and TTA thinkers use ``checkpoint_folder_audiogen``.
-        """
-        if self._audex_tokenizer is None:
-            from transformers import AutoTokenizer
-
-            from vllm_omni.model_executor.models.audex.checkpoint import ensure_audex_snapshot
-
-            stage = getattr(getattr(self._tts_stage, "engine_args", None), "model_stage", None)
-            if stage == "audex_omni":
-                profile, folder = "full", "checkpoint_folder_full"
-            elif stage == "audex_tta_thinker":
-                profile, folder = "tta", "checkpoint_folder_audiogen"
-            else:
-                profile, folder = "tts", "checkpoint_folder_audiogen"
-            # The serving model_config may already hold the RESOLVED per-stage
-            # checkpoint folder (stage init joins model_subdir onto the
-            # snapshot root); joining again would produce a non-existent path
-            # that transformers then mistakes for a repo id.
-            model_path = os.path.normpath(self.engine_client.model_config.model)
-            if os.path.basename(model_path).startswith("checkpoint_folder"):
-                root = os.path.dirname(model_path)
-            else:
-                root = ensure_audex_snapshot(model_path, profile=profile)
-            self._audex_tokenizer = AutoTokenizer.from_pretrained(os.path.join(root, folder))
-        return self._audex_tokenizer
-
-    # ---- GLM-TTS helpers ----
-
-    async def _build_glm_tts_prompt(
-        self,
-        request: OpenAICreateSpeechRequest,
-        *,
-        has_inline_ref_audio: bool = False,
-    ) -> dict[str, Any]:
-        """Build prompt for GLM-TTS.
-
-        Uses the multimodal processor path (same as CosyVoice3):
-        - prompt: synthesis text
-        - multi_modal_data["audio"]: (wav_samples, sr) reference audio
-        - mm_processor_kwargs["prompt_text"]: reference text transcript
-
-        AR preprocess() builds [PromptText | Text | BOA | PromptSpeechTokens + ATS].
-        DiT receives prompt_token, prompt_feat, embedding for conditioning.
-        """
-        # Voice cloning requires ref_audio + ref_text
-        if request.ref_audio is not None and request.ref_text:
-            wav_samples, sr, _ = await self._resolve_ref_audio(request.ref_audio)
-            audio_data = (np.asarray(wav_samples, dtype=np.float32), int(sr))
-
-            mm_kwargs: dict[str, Any] = {
-                "prompt_text": request.ref_text,
-            }
-            if request.voice:
-                voice_lower = request.voice.lower()
-                if voice_lower in self.uploaded_speakers and not has_inline_ref_audio:
-                    mm_kwargs["voice_name"] = voice_lower
-                    mm_kwargs["voice_created_at"] = self._voice_created_at(voice_lower)
-
-            return {
-                "prompt": request.input,
-                "multi_modal_data": {
-                    "audio": audio_data,
-                },
-                "mm_processor_kwargs": mm_kwargs,
-                "additional_information": self._build_glm_tts_prefill_metadata(
-                    request.input,
-                    request.ref_text,
-                ),
-            }
-
-        raise ValueError("GLM-TTS requires ref_audio and ref_text for voice cloning.")
-
-    def _glm_tts_text_tokenizer_and_frontend(self):
-        from vllm_omni.model_executor.models.glm_tts.glm_tts import (
-            load_glm_tts_tokenizer,
-            resolve_glm_tts_tokenizer_path,
-        )
-        from vllm_omni.model_executor.models.glm_tts.text_frontend import GLMTTSTextFrontend
-
-        cached = self._glm_tts_text_tokenizer
-        if cached is None:
-            model_name_or_path = self.engine_client.model_config.model
-            tokenizer_path = getattr(self.engine_client.model_config, "tokenizer", None)
-            if tokenizer_path is None:
-                tokenizer_path = resolve_glm_tts_tokenizer_path(model_name_or_path)
-            cached = load_glm_tts_tokenizer(
-                tokenizer_path,
-                model_name_or_path=model_name_or_path,
-                trust_remote_code=bool(getattr(self.engine_client.model_config, "trust_remote_code", False)),
-            )
-            self._glm_tts_text_tokenizer = cached
-
-        frontend = self._glm_tts_text_frontend
-        if frontend is None:
-            frontend = GLMTTSTextFrontend()
-            self._glm_tts_text_frontend = frontend
-        return cached, frontend
-
-    def _estimate_glm_tts_text_token_len(self, text: str | None, *, add_trailing_space: bool = False) -> int:
-        """Estimate GLM-TTS normalized text length with the model tokenizer."""
-        cached, frontend = self._glm_tts_text_tokenizer_and_frontend()
-        text = text or ""
-        normalized = frontend.text_normalize(text) or text
-        normalized = normalized.strip()
-        if add_trailing_space and normalized:
-            normalized = f"{normalized} "
-        return max(1, len(cached.encode(normalized)))
-
-    def _build_glm_tts_prefill_metadata(self, text: str, prompt_text: str | None) -> dict[str, Any]:
-        """Build GLM-TTS processor length metadata for additional_information.
-
-        The model preprocess hook runs before postprocess can mirror MM kwargs
-        into additional_information, so these scalar fields must originate from
-        the request payload rather than runner-level mm_features.
-        """
-        text_len = self._estimate_glm_tts_text_token_len(text)
-        prompt_text_len = (
-            self._estimate_glm_tts_text_token_len(prompt_text, add_trailing_space=True) if prompt_text else 0
-        )
-        return {
-            "glm_tts_text_token_len": [text_len],
-            "glm_tts_prompt_text_token_len": [prompt_text_len],
-            "input_len": [prompt_text_len + text_len + 1],
-        }
-
-    # ---- Ming-flash-omni standalone-talker (TTS) helpers ----
-
-    def _build_ming_flash_omni_prompt(self, request: OpenAICreateSpeechRequest) -> dict[str, Any]:
-        # request.instructions accepts two forms:
-        # 1. Plain text: mapped to the caption's 风格 (style) field
-        # 2. JSON object: parsed and splatted into the caption. Unlocks
-        #       Unknown keys are dropped by `ming_create_instruction`.
-        caption_fields = self._parse_ming_instruction_fields(request) or {}
-
-        has_spk_emb = request.speaker_embedding is not None
-
-        # TTS path applies ming task type `instruct`.
-        # voice_name enables talker-side voice preset resolution (e.g. "DB30").
-        additional_information: dict[str, Any] = {
-            "ming_task": "instruct",
-            "prompt": MING_DEFAULT_PROMPT,
-            "text": request.input,
-            "instruction": ming_create_instruction(caption_fields),
-            "voice_name": request.voice or None,
-            "use_zero_spk_emb": not has_spk_emb,
-            "max_decode_steps": request.max_new_tokens or _TTS_MAX_NEW_TOKENS_MAX,
-            "cfg": 2.0,
-            "sigma": 0.25,
-            "temperature": 0.0,
-        }
-        if has_spk_emb:
-            # Passed as plain float list
-            additional_information["spk_emb"] = list(request.speaker_embedding)
-        prompt = tokens_input(prompt_token_ids=[0])
-        prompt["additional_information"] = additional_information
-        return prompt
-
-    # ---- Common speech generation helpers ----
-
-    async def _build_qwen3_tts_request(
-        self,
-        request: OpenAICreateSpeechRequest,
-    ) -> tuple[dict[str, Any], dict[str, Any], str | None]:
-        """Build prompt + tts_params for Qwen3-TTS.
-
-        Called from ``Qwen3TTSAdapter.build``. Returns
-        ``(prompt, tts_params, warmup_artifact_key)`` where the warmup key is the
-        Qwen3-TTS ref-audio artifact tracked after ``generate()``.
-        """
-        qwen3_ref_audio_warmup_artifact_key: str | None = None
-        tts_params = self._build_tts_params(request)
-        # Resolve ref_audio (explicit or auto-set for uploaded voices)
-        # to [[wav_list, sr]] so the model doesn't re-decode base64.
-        ref_audio_source = request.ref_audio
-        if ref_audio_source is None and isinstance(tts_params.get("ref_audio"), list):
-            # Uploaded voice: ref_audio was auto-set as [base64_data_url]
-            ref_audio_source = tts_params["ref_audio"][0]
-        if ref_audio_source is not None and isinstance(ref_audio_source, str):
-            wav_list, sr, cache_key = await self._resolve_ref_audio(ref_audio_source)
-            tts_params["ref_audio_cache_key"] = cache_key
-            artifact_key = self._get_resolved_ref_audio_artifact_key(cache_key)
-            if artifact_key:
-                tts_params[_QWEN3_TTS_REF_AUDIO_CACHE_KEY] = [artifact_key]
-            ref_code_length = self._estimate_ref_code_len([wav_list, sr])
-            if ref_code_length is not None:
-                tts_params["ref_code_length"] = [int(ref_code_length)]
-            if self._qwen3_tts_can_use_ref_audio_artifact_only(tts_params, artifact_key):
-                logger.debug("Using Qwen3-TTS ref_audio artifact-only path: %s", artifact_key)
-            else:
-                tts_params["ref_audio"] = [[wav_list, sr]]
-                qwen3_ref_audio_warmup_artifact_key = artifact_key
-
-        ph_len = await self._estimate_prompt_len_async(tts_params)
-        prompt = tokens_input(prompt_token_ids=[1] * ph_len)
-        prompt["additional_information"] = tts_params
-        prompt["cache_salt"] = _conditioning_cache_salt(request, tts_params)
-        return prompt, tts_params, qwen3_ref_audio_warmup_artifact_key
-
     async def _prepare_speech_generation(
         self,
         request: OpenAICreateSpeechRequest,
@@ -3188,18 +1808,6 @@ class OmniOpenAIServingSpeech(OpenAIServing, AudioMixin):
             if stage0_params.extra_args is None:
                 stage0_params.extra_args = {}
             stage0_params.extra_args["tts_local_seed"] = request.seed
-
-        if self._tts_model_type == "qwen3_tts" and sampling_params_list:
-            stage0_params = sampling_params_list[0]
-            default_seed = getattr(stage0_params, "seed", None)
-            if default_seed is not None:
-                import copy
-
-                sampling_params_list = copy.deepcopy(sampling_params_list)
-                stage0_params = sampling_params_list[0]
-                if stage0_params.extra_args is None:
-                    stage0_params.extra_args = {}
-                stage0_params.extra_args.setdefault("tts_local_seed", int(default_seed))
 
         # When word_timestamps is requested, also ask for the aligner stage's
         # output so the orchestrator drives the request through the forced-aligner
