+++ vllm_omni/entrypoints/openai/tts_adapters/glm_tts.py
@@ -1,8 +1,10 @@
 # SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Omni project
 """GLM-TTS serving adapter."""
 
 from typing import TYPE_CHECKING, Any
 
+import numpy as np
 from vllm.logger import init_logger
 
 from vllm_omni.entrypoints.openai.tts_adapters import register_tts_adapter
@@ -18,6 +20,72 @@ logger = init_logger(__name__)
 class GlmTTSAdapter(ARTTSAdapter):
     stage_keys = frozenset({"glm_tts"})
     name = "glm_tts"
+
+    def __init__(self, ctx) -> None:
+        super().__init__(ctx)
+        self._text_tokenizer = None
+        self._text_frontend = None
+
+    def _text_tokenizer_and_frontend(self):
+        from vllm_omni.model_executor.models.glm_tts.glm_tts import (
+            load_glm_tts_tokenizer,
+            resolve_glm_tts_tokenizer_path,
+        )
+        from vllm_omni.model_executor.models.glm_tts.text_frontend import GLMTTSTextFrontend
+
+        if self._text_tokenizer is None:
+            config = self.ctx.engine_client.model_config
+            tokenizer_path = getattr(config, "tokenizer", None) or resolve_glm_tts_tokenizer_path(config.model)
+            self._text_tokenizer = load_glm_tts_tokenizer(
+                tokenizer_path,
+                model_name_or_path=config.model,
+                trust_remote_code=bool(getattr(config, "trust_remote_code", False)),
+            )
+        if self._text_frontend is None:
+            self._text_frontend = GLMTTSTextFrontend()
+        return self._text_tokenizer, self._text_frontend
+
+    def _estimate_text_token_len(self, text: str | None, *, add_trailing_space: bool = False) -> int:
+        tokenizer, frontend = self._text_tokenizer_and_frontend()
+        normalized = (frontend.text_normalize(text or "") or text or "").strip()
+        if add_trailing_space and normalized:
+            normalized = f"{normalized} "
+        return max(1, len(tokenizer.encode(normalized)))
+
+    def _build_prefill_metadata(self, text: str, prompt_text: str | None) -> dict[str, Any]:
+        text_len = self._estimate_text_token_len(text)
+        prompt_len = self._estimate_text_token_len(prompt_text, add_trailing_space=True) if prompt_text else 0
+        return {
+            "glm_tts_text_token_len": [text_len],
+            "glm_tts_prompt_text_token_len": [prompt_len],
+            "input_len": [prompt_len + text_len + 1],
+        }
+
+    async def _build_prompt(
+        self, request: "OpenAICreateSpeechRequest", *, has_inline_ref_audio: bool = False
+    ) -> dict[str, Any]:
+        """Build GLM-TTS multimodal voice-cloning input and prefill metadata.
+
+        AR preprocess constructs PromptText, Text, BOA and prompt speech tokens;
+        the DiT consumes the resulting prompt tokens, features and embedding as
+        conditioning.
+        """
+        server = self.ctx.server
+        if request.ref_audio is None or not request.ref_text:
+            raise ValueError("GLM-TTS requires ref_audio and ref_text for voice cloning.")
+        wav_samples, sr, _ = await server._resolve_ref_audio(request.ref_audio)
+        mm_kwargs: dict[str, Any] = {"prompt_text": request.ref_text}
+        if request.voice:
+            voice_lower = request.voice.lower()
+            if voice_lower in server.uploaded_speakers and not has_inline_ref_audio:
+                mm_kwargs["voice_name"] = voice_lower
+                mm_kwargs["voice_created_at"] = server._voice_created_at(voice_lower)
+        return {
+            "prompt": request.input,
+            "multi_modal_data": {"audio": (np.asarray(wav_samples, dtype=np.float32), int(sr))},
+            "mm_processor_kwargs": mm_kwargs,
+            "additional_information": self._build_prefill_metadata(request.input, request.ref_text),
+        }
 
     def validate(self, request: "OpenAICreateSpeechRequest") -> str | None:
         """Validate GLM-TTS request — requires ref_audio for voice cloning."""
@@ -46,7 +114,7 @@ class GlmTTSAdapter(ARTTSAdapter):
     async def build(
         self, request: "OpenAICreateSpeechRequest", sampling_params_list: list, has_inline_ref_audio: bool
     ) -> PreparedRequest:
-        prompt = await self.ctx.server._build_glm_tts_prompt(request, has_inline_ref_audio=has_inline_ref_audio)
+        prompt = await self._build_prompt(request, has_inline_ref_audio=has_inline_ref_audio)
         return PreparedRequest(prompt=prompt, tts_params={}, model_type="glm_tts")
 
     def _load_supported_speakers(self) -> set[str]:
@@ -71,9 +139,7 @@ class GlmTTSAdapter(ARTTSAdapter):
             if isinstance(text_len_value, list) and text_len_value:
                 text_len_value = text_len_value[0]
         text_token_len = (
-            int(text_len_value)
-            if text_len_value is not None
-            else server._estimate_glm_tts_text_token_len(request.input)
+            int(text_len_value) if text_len_value is not None else self._estimate_text_token_len(request.input)
         )
         hf_cfg = server.model_config.hf_config
         min_ratio = getattr(hf_cfg, "min_token_text_ratio", 2)
